% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-extensions.R
\name{assert_empty}
\alias{assert_empty}
\title{assert_empty}
\usage{
assert_empty(
  x,
  mode = typeof(x),
  null.ok = FALSE,
  zero.len.string.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode}

\item{null.ok}{allow NULL to be valid}

\item{zero.len.string.ok}{accept zero length strings}

\item{.var.name}{Name of the checked object to print in assertions. Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{Collection to store assertion messages. See \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection()}}.}
}
\value{
x or throws an error
}
\description{
assert_empty
}
\examples{

mode <- "character"
assert_empty(character(), mode)
assert_empty(character(), mode, null.ok = TRUE)
assert_empty(NULL, mode, null.ok = TRUE)
assert_empty("", zero.len.string.ok = TRUE)
assert_empty("", mode, null.ok = TRUE, zero.len.string.ok = TRUE)
}
