% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avs_roster_2021.R
\docType{data}
\name{avs_roster_2021}
\alias{avs_roster_2021}
\title{2021 NHL Colorado Avalanche Roster (from NHL's API)}
\format{
A data frame with 28 rows and 8 variables:
\describe{
  \item{jersey_number}{player jersey number}
  \item{person_id}{player ID}
  \item{person_full_name}{player full name}
  \item{person_link}{API link to player data}
  \item{position_code}{position abbreviation}
  \item{position_name}{position name}
  \item{position_type}{position type}
  \item{position_abbreviation}{another position abbreviation}
}
}
\source{
\url{https://www.nhl.com/stats/}
}
\usage{
avs_roster_2021
}
\description{
A dataset containing the NHL's Avalanche Roster
}
\keyword{datasets}
