% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avs_stats_2021.R
\docType{data}
\name{avs_stats_2021}
\alias{avs_stats_2021}
\title{NHL's Colorado Avalanche Player Statistics}
\format{
A data frame with 196 rows and 50 variables:
\describe{
  \item{person_id}{Season in which the game was played (YYYY)}
  \item{person_full_name}{Season in which the game was played (YYYY-YY)}
  \item{season}{Always given as NBA}
  \item{season_start}{Regular season data}
  \item{stat_assists}{Date in YYYY-MM-DD}
  \item{stat_goals}{Game ID on nba.com}
  \item{stat_pim}{slugTeam game number in the season}
  \item{stat_games}{Full team name}
  \item{stat_penalty_minutes}{Team ID on nba.com}
  \item{stat_plus_minus}{Logical for back-to-back game}
  \item{stat_points}{Logical for first game of back-to-back game}
  \item{stat_power_play_goals}{Logical for second game of back-to-back game}
  \item{stat_game_winning_goals}{Home (H) or Away (A)}
  \item{stat_short_handed_goals}{Team One vs. Team Two}
  \item{stat_shots}{Team abbreviation}
  \item{stat_shot_pct}{Days between previous game and this game}
  \item{stat_time_on_ice}{Days until next game}
  \item{stat_hits}{Opponent abbreviation}
  \item{stat_power_play_points}{Team that won (abbreviation)}
  \item{stat_power_play_time_on_ice}{Team that lost (abbreviation)}
  \item{stat_even_time_on_ice}{Result for slugTeam (W or L)}
  \item{stat_face_off_pct}{Logical for slugTeam winning}
  \item{stat_over_time_goals}{slugTeam field goals made}
  \item{stat_short_handed_points}{slugTeam field goal attempts}
  \item{stat_blocked}{slugTeam field goal percentage}
  \item{stat_shifts}{slugTeam three point field goals made}
  \item{stat_save_percentage}{slugTeam three point field goal attempts}
  \item{stat_goal_against_average}{slugTeam three point field goal percentage}
  \item{stat_ot}{slugTeam free throw percentage}
  \item{stat_shutouts}{Logical if game video exists}
  \item{stat_ties}{slugTeam two point field goals made}
  \item{stat_wins}{slugTeam two point field goal attempts}
  \item{stat_losses}{slugTeam two point field goal percentage}
  \item{stat_saves}{Total minutes played}
  \item{stat_power_play_saves}{slugTeam free throws made}
  \item{stat_short_handed_saves}{slugTeam free throw attempts}
  \item{stat_even_saves}{slugTeam offensive rebounds}
  \item{stat_short_handed_shots}{slugTeam defensive rebounds}
  \item{stat_even_shots}{slugTeam total rebounds}
  \item{stat_power_play_shots}{slugTeam total assists}
  \item{stat_games_started}{slugTeam total steals}
  \item{stat_shots_against}{slugTeam total blocks}
  \item{stat_goals_against}{slugTeam total turnovers}
  \item{stat_power_play_save_percentage}{slugTeam total personal fouls}
  \item{stat_short_handed_save_percentage}{slugTeam total points scored}
  \item{stat_even_strength_save_percentage}{Margin of game as slugTeam score minus slugOpponent score}
  \item{team_name}{url of slugTeam logo}
  \item{team_link}{Margin of game as slugTeam score minus slugOpponent score}
  \item{team_id}{url of slugTeam logo}
}
}
\source{
\url{https://statsapi.web.nhl.com/api/v1}
}
\usage{
avs_stats_2021
}
\description{
A dataset containing the statistics for all Colorado Avalanche players on the 2021-2 roster
}
\keyword{datasets}
