% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characteristic.R
\name{Characteristic}
\alias{Characteristic}
\title{A Characteristic acts as a qualifying property to a material object.}
\description{
A Characteristic acts as a qualifying property to a material object.

A Characteristic acts as a qualifying property to a material object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{category}}{The classifier of the type of characteristic being described.}

\item{\code{value}}{The value of this instance of a characteristic as relevant to the attached material.}

\item{\code{unit}}{If applicable, a unit qualifier for the value (if the value is numeric).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Characteristic-new}{\code{Characteristic$new()}}
\item \href{#method-Characteristic-check_category}{\code{Characteristic$check_category()}}
\item \href{#method-Characteristic-check_value}{\code{Characteristic$check_value()}}
\item \href{#method-Characteristic-check_unit}{\code{Characteristic$check_unit()}}
\item \href{#method-Characteristic-set_category}{\code{Characteristic$set_category()}}
\item \href{#method-Characteristic-set_value}{\code{Characteristic$set_value()}}
\item \href{#method-Characteristic-set_unit}{\code{Characteristic$set_unit()}}
\item \href{#method-Characteristic-to_list}{\code{Characteristic$to_list()}}
\item \href{#method-Characteristic-from_list}{\code{Characteristic$from_list()}}
\item \href{#method-Characteristic-get_id}{\code{Characteristic$get_id()}}
\item \href{#method-Characteristic-clone}{\code{Characteristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$new(category = NULL, value = NULL, unit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_category}{}}}
\subsection{Method \code{check_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_category()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_value}{}}}
\subsection{Method \code{check_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_unit}{}}}
\subsection{Method \code{check_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_unit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_category}{}}}
\subsection{Method \code{set_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_category()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_unit}{}}}
\subsection{Method \code{set_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_unit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$from_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
