% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacteristicCategory.R
\name{CharacteristicCategory}
\alias{CharacteristicCategory}
\title{CharacteristicCategory}
\description{
CharacteristicCategory

CharacteristicCategory
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{@id}}{identifier}

\item{\code{type}}{The type of characteristic as an \link{OntologyAnnotation}}

\item{\code{explicitly_provided}}{boolean was this characteristic explicitly as a characteristic category of this study (deprecating in favour of source)}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}

\item{\code{source}}{an id of the object from which characteristic category originated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CharacteristicCategory-new}{\code{CharacteristicCategory$new()}}
\item \href{#method-CharacteristicCategory-check_id}{\code{CharacteristicCategory$check_id()}}
\item \href{#method-CharacteristicCategory-set_id}{\code{CharacteristicCategory$set_id()}}
\item \href{#method-CharacteristicCategory-check_type}{\code{CharacteristicCategory$check_type()}}
\item \href{#method-CharacteristicCategory-set_type}{\code{CharacteristicCategory$set_type()}}
\item \href{#method-CharacteristicCategory-to_list}{\code{CharacteristicCategory$to_list()}}
\item \href{#method-CharacteristicCategory-from_list}{\code{CharacteristicCategory$from_list()}}
\item \href{#method-CharacteristicCategory-print}{\code{CharacteristicCategory$print()}}
\item \href{#method-CharacteristicCategory-clone}{\code{CharacteristicCategory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-new"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$new(
  `@id` = character(),
  type = NULL,
  explicitly_provided = logical(),
  source = character(),
  ontology_source_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{@id}}{identifier}

\item{\code{type}}{The type of characteristic as an \link{OntologyAnnotation}}

\item{\code{explicitly_provided}}{boolean was this characteristic explicitly as a characteristic category of this study (deprecating in favour of source)}

\item{\code{source}}{an id of the object from which characteristic category originated}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{CharacteristicCategory} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-check_id"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-check_id}{}}}
\subsection{Method \code{check_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$check_id(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the identifier of the object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that @id is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$set_id(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the identifier of the object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the object @id if it is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-check_type"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-check_type}{}}}
\subsection{Method \code{check_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$check_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{the type of of the characteristic}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the type is an \link{OntologyAnnotation} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-set_type"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-set_type}{}}}
\subsection{Method \code{set_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$set_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{the type of of the characteristic}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the type if the type is an \link{OntologyAnnotion} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
seralise the \link{CharacteristicCategory} object to an R list
}

\subsection{Returns}{
an R list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Populate a \link{CharacteristicCategory} object from a list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-print"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{CharacteristicCategory} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
