% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharacteristicCategoryReferences.R
\name{CharacteristicCategoryReferences}
\alias{CharacteristicCategoryReferences}
\title{CharacteristicCategoryReferences}
\description{
CharacteristicCategoryReferences

CharacteristicCategoryReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{categories}}{A list of \link{CharacteristicCategory} objects}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object listing all the ontology sources used}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CharacteristicCategoryReferences-new}{\code{CharacteristicCategoryReferences$new()}}
\item \href{#method-CharacteristicCategoryReferences-check_categories}{\code{CharacteristicCategoryReferences$check_categories()}}
\item \href{#method-CharacteristicCategoryReferences-add_categories}{\code{CharacteristicCategoryReferences$add_categories()}}
\item \href{#method-CharacteristicCategoryReferences-set_categories}{\code{CharacteristicCategoryReferences$set_categories()}}
\item \href{#method-CharacteristicCategoryReferences-get_characterisic_categories}{\code{CharacteristicCategoryReferences$get_characterisic_categories()}}
\item \href{#method-CharacteristicCategoryReferences-check_type}{\code{CharacteristicCategoryReferences$check_type()}}
\item \href{#method-CharacteristicCategoryReferences-set_type}{\code{CharacteristicCategoryReferences$set_type()}}
\item \href{#method-CharacteristicCategoryReferences-get_category_ids}{\code{CharacteristicCategoryReferences$get_category_ids()}}
\item \href{#method-CharacteristicCategoryReferences-get_category_names}{\code{CharacteristicCategoryReferences$get_category_names()}}
\item \href{#method-CharacteristicCategoryReferences-get_characteristic_category_origins}{\code{CharacteristicCategoryReferences$get_characteristic_category_origins()}}
\item \href{#method-CharacteristicCategoryReferences-to_list}{\code{CharacteristicCategoryReferences$to_list()}}
\item \href{#method-CharacteristicCategoryReferences-from_list}{\code{CharacteristicCategoryReferences$from_list()}}
\item \href{#method-CharacteristicCategoryReferences-print}{\code{CharacteristicCategoryReferences$print()}}
\item \href{#method-CharacteristicCategoryReferences-clone}{\code{CharacteristicCategoryReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$new(
  categories = list(),
  ontology_source_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{characteristic categories}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object listing all the ontology sources used}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{CharacteristicCategoryReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-check_categories"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-check_categories}{}}}
\subsection{Method \code{check_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$check_categories(categories)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{characteristic categories}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if categories are \link{CharacteristicCategory} references
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-add_categories"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-add_categories}{}}}
\subsection{Method \code{add_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$add_categories(categories)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{characteristic categories}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add categories to the reference an check that they are
\link{CharacteristicCategory} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-set_categories"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-set_categories}{}}}
\subsection{Method \code{set_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$set_categories(categories)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{categories}}{Characterisic categories}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the categories in the characteristic category reference,
overwrites existing categories
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-get_characterisic_categories"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-get_characterisic_categories}{}}}
\subsection{Method \code{get_characterisic_categories()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$get_characterisic_categories(names = "all")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{default return all names}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get characteristic categories
}

\subsection{Returns}{
a vector of \link{CharacteristicCategory} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-check_type"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-check_type}{}}}
\subsection{Method \code{check_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$check_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{category type}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Checks that the type is \link{CharacteristicCategory}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-set_type"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-set_type}{}}}
\subsection{Method \code{set_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$set_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{category type}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the type value
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-get_category_ids"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-get_category_ids}{}}}
\subsection{Method \code{get_category_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$get_category_ids()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the ids of the categories in the reference
}

\subsection{Returns}{
character vector of category names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-get_category_names"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-get_category_names}{}}}
\subsection{Method \code{get_category_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$get_category_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the names of the categories in the reference
}

\subsection{Returns}{
character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-get_characteristic_category_origins"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-get_characteristic_category_origins}{}}}
\subsection{Method \code{get_characteristic_category_origins()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$get_characteristic_category_origins()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the source of the characteristic category, such as from which
study does it originate
}

\subsection{Returns}{
character vector of characteristic category sources
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$to_list(origin = "any")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{origin}}{ids of sources of characteristic categories for which
to generate a list of category references.
"any" lists categories from all sources.
default = "any"}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Serialize \link{CharacteristicCategoryReferences} object to an R list
}

\subsection{Returns}{
an R list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$from_list(lst, origin = NA, add = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list from which to generate a \link{CharacteristicCategoryReferences} object}

\item{\code{origin}}{the id of the object in which this category originated}

\item{\code{add}}{(logical) if true append new charateristic categoeries to
the ones already present in the reference, if fale then overwrite and
existing ones.}

\item{\code{explicitly_provided}}{(logical) if true the category was
explicitly declared if false it was generated automatically as no
explicitly declared category was found}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Populate \link{CharacteristicCategoryReferences} object from a list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-print"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{CharacteristicCategoryReferences} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CharacteristicCategoryReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CharacteristicCategoryReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CharacteristicCategoryReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
