% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_comment}
\alias{check_comment}
\title{check_comment}
\usage{
check_comment(comment)
}
\arguments{
\item{comment}{A comment
(a list with names 'name' and 'value' each length 1 character vectors)}
}
\description{
check_comment
}
