% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_comments}
\alias{check_comments}
\title{check_comments}
\usage{
check_comments(comments)
}
\arguments{
\item{comments}{a list of comments, named and with length 1 character vectors#}
}
\description{
check_comments
}
