% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-extensions.R
\name{check_empty}
\alias{check_empty}
\title{check_empty}
\usage{
check_empty(x, mode = typeof(x), null.ok = FALSE, zero.len.string.ok = FALSE)
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode, defaults to the type of the vector}

\item{null.ok}{allow NULL to be valid}

\item{zero.len.string.ok}{accept zero length strings}
}
\value{
TRUE or string containing an error message
}
\description{
check_empty
}
\examples{

mode <- "character"
check_empty(character(), mode)
check_empty(character(), mode, null.ok = TRUE)
check_empty(NULL, mode, null.ok = TRUE)
check_empty("", mode)
check_empty("", mode, null.ok = TRUE)

}
