% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_id}
\alias{check_id}
\title{check_id
Checks if an id is valid, returns TRUE if valid and explanatory string if not
(follows checkmate conventions)}
\usage{
check_id(id)
}
\arguments{
\item{id}{a uuid or uuid with a suffix}
}
\description{
check_id
Checks if an id is valid, returns TRUE if valid and explanatory string if not
(follows checkmate conventions)
}
