% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{check_isa_tab_dir}
\alias{check_isa_tab_dir}
\title{check_isa_tab_dir}
\usage{
check_isa_tab_dir(path)
}
\arguments{
\item{path}{the path to an ISA-tab directory}
}
\value{
a string or logical
}
\description{
Checks if a directory exists and contains the right files for an ISA-tab
description.
}
\details{
Follows the pattern of the \verb{check_*} functions from the
\href{https://mllg.github.io/checkmate/}{checkmate} package.
}
