% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySourceReferences.R
\name{check_ontology_source_references}
\alias{check_ontology_source_references}
\title{check_ontology_source_references}
\usage{
check_ontology_source_references(ontology_source_references)
}
\arguments{
\item{ontology_source_references}{something you want to check is an
\link{OntologySourceReferences} object.}
}
\description{
returns TRUE if ontology_source_references is an \link{OntologySourceReferences}
object and throws an error if it is not
}
