% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolReferences.R
\name{check_protocol_references}
\alias{check_protocol_references}
\title{check_protocol_references}
\usage{
check_protocol_references(protocol_references)
}
\arguments{
\item{protocol_references}{something you want to check is an
\link{ProtocolReferences} object.}
}
\description{
returns TRUE if protocol_references is an \link{ProtocolReferences}
object and throws an error if it is not
}
