% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitReferences.R
\name{check_unit_references}
\alias{check_unit_references}
\title{check_unit_references}
\usage{
check_unit_references(unit_references)
}
\arguments{
\item{unit_references}{something you want to check is an
\link{UnitReferences} object.}
}
\description{
returns TRUE if unit_references is an \link{UnitReferences}
object and throws an error if it is not
}
