% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{clear_all_na_cols}
\alias{clear_all_na_cols}
\title{clear_all_na_cols}
\usage{
clear_all_na_cols(tbl)
}
\arguments{
\item{tbl}{a tibble}
}
\value{
a tibble
}
\description{
Removes columns the contents of which are all NA
}
\examples{

clear_all_na_cols(tibble::tibble(A = c(NA, NA), B = 1:2, C = c(NA, FALSE)))

}
