% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{clear_na_cols_and_transpose}
\alias{clear_na_cols_and_transpose}
\title{clear_na_cols_and_transpose}
\usage{
clear_na_cols_and_transpose(tbl)
}
\arguments{
\item{tbl}{a tibble}
}
\value{
a tibble
}
\description{
convenience wrapper for doing first removal of columns with all NA values
and transposition of a tibble where first row becomes the column names
}
\examples{
clear_na_cols_and_transpose(tibble::tibble(A = c(NA, NA), B = 1:2, C = c(NA, FALSE)))

}
