% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comment.R
\name{Comment}
\alias{Comment}
\title{R6 class for comments}
\description{
A comment allows arbitrary annotation of all commentable ISA classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{a name for the comment string}

\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Comment-new}{\code{Comment$new()}}
\item \href{#method-Comment-check_name}{\code{Comment$check_name()}}
\item \href{#method-Comment-check_value}{\code{Comment$check_value()}}
\item \href{#method-Comment-set_name}{\code{Comment$set_name()}}
\item \href{#method-Comment-set_value}{\code{Comment$set_value()}}
\item \href{#method-Comment-to_list}{\code{Comment$to_list()}}
\item \href{#method-Comment-from_list}{\code{Comment$from_list()}}
\item \href{#method-Comment-clone}{\code{Comment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$new(name = NULL, value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}

\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make a new comment object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that name is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-check_value"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-check_value}{}}}
\subsection{Method \code{check_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$check_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that value is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set a valid name
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set a valid value
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
convert object to an R list
}

\subsection{Returns}{
a list of length 1 where name is the name and value the value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$from_list(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of length 1 where the name is the name and the value the value of the comment}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate a comment object from an R list
}

\subsection{Returns}{
a list of length 1 where name is the name and value the value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Comment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Comment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Comment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
