% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentable.R
\name{commentable}
\alias{commentable}
\title{R6 superclass for making classes commentable}
\description{
R6 superclass for making classes commentable

R6 superclass for making classes commentable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-commentable-new}{\code{commentable$new()}}
\item \href{#method-commentable-check_comments}{\code{commentable$check_comments()}}
\item \href{#method-commentable-set_comments}{\code{commentable$set_comments()}}
\item \href{#method-commentable-get_comments}{\code{commentable$get_comments()}}
\item \href{#method-commentable-get_comment}{\code{commentable$get_comment()}}
\item \href{#method-commentable-get_comment_names}{\code{commentable$get_comment_names()}}
\item \href{#method-commentable-get_comment_values}{\code{commentable$get_comment_values()}}
\item \href{#method-commentable-add_comment}{\code{commentable$add_comment()}}
\item \href{#method-commentable-to_list}{\code{commentable$to_list()}}
\item \href{#method-commentable-from_list}{\code{commentable$from_list()}}
\item \href{#method-commentable-clone}{\code{commentable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-new"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$new(comments = list(comment$new()))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new \code{[commentable]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check comments is a list of \code{[comment]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set comments if comments is a list of \code{[comment]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-get_comments"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-get_comments}{}}}
\subsection{Method \code{get_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$get_comments()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get comments as a list of \code{[comment]} objects
}

\subsection{Returns}{
comments a list of \code{[comment]} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-get_comment"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-get_comment}{}}}
\subsection{Method \code{get_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$get_comment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the comment to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
get a specific comment by name as a \code{[comment]} object
}

\subsection{Returns}{
selected comment a \code{[comment]} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-get_comment_names"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-get_comment_names}{}}}
\subsection{Method \code{get_comment_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$get_comment_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the names of all comments
}

\subsection{Returns}{
a character vector of comment names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-get_comment_values"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-get_comment_values}{}}}
\subsection{Method \code{get_comment_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$get_comment_values()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the values of all comments
}

\subsection{Returns}{
a character vector of comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a \code{[comment]} object to append to the existing \code{[commentable]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
append a new \code{[comment]} object to the existing list of comments
}

\subsection{Returns}{
a character vector of comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
convert \code{[commentable]} object into an R list of comments
}

\subsection{Returns}{
a list where the names are comment names and the values comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$from_list(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a R list where the names are comment names and the values are comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
convert an R list into a object \code{[commentable]}
}

\subsection{Returns}{
commentable a \code{[commentable]} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commentable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-commentable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{commentable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
