% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRediT}
\alias{CRediT}
\title{CRediT taxonomy}
\format{
\link{OntologySource} object
}
\source{
\url{https://credit.niso.org/}
}
\usage{
CRediT
}
\description{
CRediT taxonomy
}
\keyword{datasets}
