% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_investigation}
\alias{credit_investigation}
\title{CRediT taxonomy: Investigation}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/investigation/}
}
\usage{
credit_investigation
}
\description{
CRediT taxonomy: Investigation
}
\keyword{datasets}
