% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_project_administration}
\alias{credit_project_administration}
\title{CRediT taxonomy: Project administration}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/project-administration/}
}
\usage{
credit_project_administration
}
\description{
CRediT taxonomy: Project administration
}
\keyword{datasets}
