% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_resources}
\alias{credit_resources}
\title{CRediT taxonomy: Resources}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/resources/}
}
\usage{
credit_resources
}
\description{
CRediT taxonomy: Resources
}
\keyword{datasets}
