% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data curation}
\alias{Data curation}
\title{CRediT taxonomy: Data curation}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/data-curation/}
}
\usage{
Data curation
}
\description{
CRediT taxonomy: Data curation
}
\keyword{datasets}
