% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{date_input_handling}
\alias{date_input_handling}
\title{date_input_handling}
\usage{
date_input_handling(date, strict = TRUE, null.ok = FALSE)
}
\arguments{
\item{date_string}{a string to be converted to a date}
}
\value{
a Date object
}
\description{
date_input_handling
}
\examples{
date_input_handling("2012-08-01")
date_input_handling("1Jan2009", strict = FALSE)


}
