% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{date_string_conversion}
\alias{date_string_conversion}
\title{date_string_conversion}
\usage{
date_string_conversion(date_string)
}
\arguments{
\item{date_string}{a string to be converted to a date}
}
\value{
a Date object
}
\description{
date_string_conversion
}
\examples{
date_string_conversion("2012-08-01")
date_string_conversion("1Jan2009")


}
