% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings-bos-buf.R
\docType{data}
\name{dk_bos_buf}
\alias{dk_bos_buf}
\title{Draftkings Showdown Game in NHL (Boston Bruins at Buffalo Sabres, April 21, 2026)}
\format{
A data frame with 128 rows and 9 variables:
\describe{
  \item{position}{real-life roster position of the player}
  \item{name_id}{player name and Draftkings ID}
  \item{name}{player name}
  \item{id}{Draftkings player ID}
  \item{roster_position}{Draftkings roster position}
  \item{salary}{player salary}
  \item{game_info}{game information for the players}
  \item{team_abbrev}{player team as an abbreviation}
  \item{avg_points_per_game}{player average Draftkings points per game}
}
}
\source{
Draftkings Contest Page
}
\usage{
dk_bos_buf
}
\description{
A dataset containing the Draftkings salary information for Game 2 of the
 2026 NHL Playoffs between the Boston Bruins and Buffalo Sabres (round 1).
}
\keyword{datasets}
