% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draftkings-lac-dal.R
\docType{data}
\name{dk_lac_dal}
\alias{dk_lac_dal}
\title{Draftkings Showdown Game (Clippers vs. Mavericks)}
\format{
A data frame with 68 rows and 9 variables:
\describe{
  \item{position}{real-life roster position of the player}
  \item{name_id}{player name and Draftkings ID}
  \item{name}{player name}
  \item{id}{Draftkings player ID}
  \item{roster_position}{Draftkings roster position}
  \item{salary}{player salary}
  \item{game_info}{game information for the players}
  \item{team_abbrev}{player team as an abbreviation}
  \item{avg_points_per_game}{player average Draftkings points per game}
}
}
\source{
Draftkings Contest Page
}
\usage{
dk_lac_dal
}
\description{
A dataset containing the Draftkings salary information for Game 3 of the
 2021 NBA Playoffs between the Los Angeles Clippers and the Dallas Mavericks.
}
\keyword{datasets}
