% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{enumerate_all_roles}
\alias{enumerate_all_roles}
\title{enumerate_all_roles}
\usage{
enumerate_all_roles(isa_json)
}
\arguments{
\item{isa_json}{isa json read in by jsonlite}
}
\value{
a list of roles in the form of a 3 element list with names
termAccession termSource annotationValue, the components of an OntologyAnnotation
}
\description{
enumerate_all_roles
}
