% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{enumerate_roles}
\alias{enumerate_roles}
\title{enumerate_roles}
\usage{
enumerate_roles(persons)
}
\arguments{
\item{persons}{list of persons from isa json read in by jsonlite}
}
\value{
a list of roles in the form of a 3 element list with names
termAccession termSource annotationValue, the components of an OntologyAnnotation
}
\description{
enumerate_roles
}
