% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl_player_stats_2022.R
\docType{data}
\name{epl_player_stats_2022}
\alias{epl_player_stats_2022}
\title{2021-2022 English Premier League Player Statistics}
\format{
A data frame with 314 rows and 13 variables:
\describe{
  \item{player}{Player name}
  \item{nation}{Player home country}
  \item{pos}{Player position}
  \item{squad}{Player team}
  \item{age}{Age of player}
  \item{born}{Birth year of player}
  \item{mp}{Matches played}
  \item{starts}{Number of matches in which player started the game}
  \item{min}{Total minutes played in the season}
  \item{min_p90}{Total minutes played in the season per ninety minutes}
  \item{goals}{Goals scored by player}
  \item{assists}{Assists by player}
  \item{goals_assists}{Goals scored plus assists by player}
  \item{non_pen_goals}{Non-penalty goals scored by player}
  \item{pen_goals}{Penalty kick goals scored by player}
  \item{pen_att}{Penalty kick attempts by player}
  \item{yellow_cards}{Yellow cards earned by player}
  \item{red_cards}{Red cards earned by player}
  \item{xG}{Expected goals}
  \item{non_pen_xG}{Non-penalty kick expected goals}
  \item{xA}{Expected assists}
  \item{non_pen_xGxA}{Non-penalty kick expected goals and assists}
  \item{prg_carry}{Progressive carries}
  \item{prg_pass}{Progressive passes}
  \item{prg_reception}{Progressive passes received}
  \item{goals_p90}{Goals per ninety minutes}
  \item{assists_p90}{Assists per ninety minutes}
  \item{goals_assists_p90}{Goals and assists per ninety minutes}
  \item{non_pen_goals_p90}{Non-penalty kick goals per ninety minutes}
  \item{non_pen_goals_assists_p90}{Non-penalty kick goals and assists per ninety minutes}
  \item{xG_p90}{Expected goals per ninety minutes}
  \item{xA_p90}{Expected assists per ninety minutes}
  \item{xG_xA_p90}{Expected goals plus assists per ninety minutes}
  \item{non_pen_xG_p90}{Expected goals minus penalty goals per ninety minutes}
  \item{non_pen_xG_xA_p90}{Expected goals plus assists minus penalty goals per ninety minutes}
}
}
\source{
\url{https://fbref.com/en/comps/9/2021-2022/stats/2021-2022-Premier-League-Stats}
}
\usage{
epl_player_stats_2022
}
\description{
A dataset containing individual player statistics for the English Premier League's 2021 - 2022 season
}
\keyword{datasets}
