% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-extensions.R
\name{expect_empty}
\alias{expect_empty}
\title{expect_empty}
\usage{
expect_empty(
  x,
  mode = typeof(x),
  null.ok = FALSE,
  zero.len.string.ok = FALSE,
  info = NULL,
  label = vname(x)
)
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode}

\item{null.ok}{allow NULL to be valid}

\item{zero.len.string.ok}{accept zero length strings}

\item{info}{default = NULL Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat:expect_that]{testthat::expect_that()}}.}

\item{label}{default = vname(x) Name of the checked object to print in messages.
Defaults to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}
}
\value{
use with testthat
}
\description{
expect_empty
}
\examples{

mode <- "character"
expect_empty(character(), mode)
expect_empty(character(), mode, null.ok = TRUE)
expect_empty(NULL, mode, null.ok = TRUE)
expect_empty("", mode, zero.len.string.ok = TRUE)
expect_empty("", mode, null.ok = TRUE, zero.len.string.ok = TRUE)

}
