% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_value.R
\name{factor_value}
\alias{factor_value}
\title{R6 class for an experimental factor value}
\description{
R6 class for an experimental factor value

R6 class for an experimental factor value
}
\details{
A factor_value represents the value instance of a \link{study_factor}.
}
\section{Public fields}{

\itemize{
\item \code{factor_name}: name of a quantity
\item \code{value}: value of a quantity
\item \code{unit}: units in which that quantity is measured
\item \code{comments}: comments
}
}

\examples{

## ------------------------------------------------
## Method `factor_value$new`
## ------------------------------------------------

factor_value$new(
    factor_name = "height",
    value = "180", unit = "cm"
)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-factor_value-new}{\code{factor_value$new()}}
\item \href{#method-factor_value-clone}{\code{factor_value$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-factor_value-new"></a>}}
\if{latex}{\out{\hypertarget{method-factor_value-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{factor_value$new(factor_name = "", value = "", unit = "", comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_name}}{the name of the experimental factor}

\item{\code{value}}{the value of a quantity}

\item{\code{unit}}{the unit of measurent}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new factor value
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{factor_value$new(
    factor_name = "height",
    value = "180", unit = "cm"
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-factor_value-clone"></a>}}
\if{latex}{\out{\hypertarget{method-factor_value-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{factor_value$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
