% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorValue.R
\name{FactorValue}
\alias{FactorValue}
\title{R6 class for an experimental factor value}
\description{
R6 class for an experimental factor value

R6 class for an experimental factor value
}
\details{
A factor_value represents the value instance of a \link{StudyFactor}.
}
\examples{

## ------------------------------------------------
## Method `FactorValue$new`
## ------------------------------------------------

\dontrun{
FactorValue$new(
    factor_name = sf,
    value = "agrees with",
    unit = NULL
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{factor_name}}{name of a quantity}

\item{\code{value}}{value of a quantity}

\item{\code{unit}}{units in which that quantity is measured}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorValue-new}{\code{FactorValue$new()}}
\item \href{#method-FactorValue-to_list}{\code{FactorValue$to_list()}}
\item \href{#method-FactorValue-from_list}{\code{FactorValue$from_list()}}
\item \href{#method-FactorValue-clone}{\code{FactorValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$new(factor_name = NULL, value = NULL, unit = NULL, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_name}}{the name of the experimental factor}

\item{\code{value}}{the value of a quantity}

\item{\code{unit}}{the unit of measurement}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new factor value
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
FactorValue$new(
    factor_name = sf,
    value = "agrees with",
    unit = NULL
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{OntologyAnnotation} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
