% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formal_analysis}
\alias{formal_analysis}
\title{CRediT taxonomy: Formal analysis}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/formal-analysis/}
}
\usage{
formal_analysis
}
\description{
CRediT taxonomy: Formal analysis
}
\keyword{datasets}
