% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate_id}
\alias{generate_id}
\title{generate_id
generates a uuid with optional suffix}
\usage{
generate_id(id = uuid::UUIDgenerate(), suffix = character())
}
\arguments{
\item{id}{a uuid}

\item{suffix}{a human readable string to be suffixed to the uuid}
}
\description{
generate_id
generates a uuid with optional suffix
}
