% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{generate_ontology_source_for_roles_without_one_defined}
\alias{generate_ontology_source_for_roles_without_one_defined}
\title{generate_ontology_source_for_roles_without_one_defined}
\usage{
generate_ontology_source_for_roles_without_one_defined(roles)
}
\arguments{
\item{roles}{list of roles}
}
\value{
list of \link{OntologySource} objects or NULL
}
\description{
generate_ontology_source_for_roles_without_one_defined
}
