% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{generate_ontology_source_from_annotation_with_undefined_source}
\alias{generate_ontology_source_from_annotation_with_undefined_source}
\title{generate_ontology_source_from_annotation_with_undefined_source}
\usage{
generate_ontology_source_from_annotation_with_undefined_source(
  annotation_list,
  source_name = "UndefinedSource",
  source_description = "undefined source",
  ...
)
}
\arguments{
\item{annotation_list}{list of 3 elements lists with the names
termAccession termSource annotationValue}

\item{source_name}{The name for the \link{OntologySource} DEFAULT UndefinedSource}

\item{source_description}{The description of the \link{OntologySource} DEFAULT undefined source}

\item{...}{other arguments to \code{OntologySource$new()}}
}
\value{
an \link{OntologySource} object or NULL if there are no undefined sources
}
\description{
takes a list of ontology annotation information in the form of
a 3 element list with names termAccession termSource annotationValue,
the components of an OntologyAnnotation but with the source undefined.
Creates a placeholder \link{OntologySource} object for all the annotationValues
in this list where the source is undefined. Accesssion is ignored as it
is meaningless in the absence of a source
}
\details{
If the input items with undefined \link{OntologySource}s are of the same type
this provides a quick way to make a new OntologySource object for them
}
