% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{get_isatab_files}
\alias{get_isatab_files}
\title{get_isatab_files}
\usage{
get_isatab_files(path, verbose = FALSE)
}
\arguments{
\item{path}{path to the directory containing isatab files}

\item{verbose}{produce a message listing the number of files of each type
found (Default: FALSE)}
}
\value{
A tibble with the columns paths, files, type.
Where path is the relative path, file is the filename, and type i,s, or a
}
\description{
Generates a table listing the files of an ISA-tab dataset from the directory
which contains them.
}
