% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{get_process_sequence_order_from_json}
\alias{get_process_sequence_order_from_json}
\title{get_process_sequence_order_from_json}
\usage{
get_process_sequence_order_from_json(process_sequence)
}
\arguments{
\item{process_sequence}{}
}
\value{
an integer vector
}
\description{
get_process_sequence_order_from_json
}
