% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_r6_class}
\alias{get_r6_class}
\title{get_r6_class}
\usage{
get_r6_class(x)
}
\arguments{
\item{x}{an R6 object for which to get the class}
}
\description{
Checks if an object is an R6 object and if it is returns the class.
An error is thrown if the object is not and R6 object.
}
