% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{green_bold_name_plain_content}
\alias{green_bold_name_plain_content}
\title{green_bold_name_plain_content}
\usage{
green_bold_name_plain_content(label, content)
}
\arguments{
\item{label}{key}

\item{content}{value}
}
\description{
make text green and bold with cli
}
