% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay.R
\name{identical.Assay}
\alias{identical.Assay}
\title{identical.Assay}
\usage{
identical.Assay(x, y)
}
\arguments{
\item{x}{a \code{[Assay]} object}

\item{y}{a \code{[Assay]} object}
}
\description{
Allows checking for the identity of \code{[Assay]} objects
}
