% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characteristic.R
\name{identical.Characteristic}
\alias{identical.Characteristic}
\title{identical.Characteristic}
\usage{
identical.Characteristic(x, y)
}
\arguments{
\item{x}{a \code{[Characteristic]} object}

\item{y}{a \code{[Characteristic]} object}
}
\description{
Allows checking for the identity of \code{[Characteristic]} objects
}
