% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFile.R
\name{identical.DataFile}
\alias{identical.DataFile}
\title{identical.DataFile}
\usage{
identical.DataFile(x, y)
}
\arguments{
\item{x}{a \code{[DataFile]} object}

\item{y}{a \code{[DataFile]} object}
}
\description{
Allows checking for the identity of \code{[DataFile]} objects
}
