% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorValue.R
\name{identical.FactorValue}
\alias{identical.FactorValue}
\title{identical.FactorValue}
\usage{
identical.FactorValue(x, y)
}
\arguments{
\item{x}{a \code{[FactorValue]} object}

\item{y}{a \code{[FactorValue]} object}
}
\description{
Allows checking for the identity of \code{[FactorValue]} objects
}
