% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investigation.R
\name{identical.Investigation}
\alias{identical.Investigation}
\title{identical.Investigation}
\usage{
identical.Investigation(x, y)
}
\arguments{
\item{x}{a \link{Investigation} object}

\item{y}{a \link{Investigation} object}
}
\description{
Allows checking for the identity of \link{Investigation} objects
}
