% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Material.R
\name{identical.Material}
\alias{identical.Material}
\title{identical.Material}
\usage{
identical.Material(x, y)
}
\arguments{
\item{x}{a \code{[Material]} object}

\item{y}{a \code{[Material]} object}
}
\description{
Allows checking for the identity of \code{[Material]} objects
}
