% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyAnnotation.R
\name{identical.OntologyAnnotation}
\alias{identical.OntologyAnnotation}
\title{identical.OntologyAnnotation}
\usage{
identical.OntologyAnnotation(x, y)
}
\arguments{
\item{x}{a \code{[OntologyAnnotation]} object}

\item{y}{a \code{[OntologyAnnotation]} object}
}
\description{
Allows checking for the identity of \code{[OntologyAnnotation]} objects
}
