% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySource.R
\name{identical.OntologySource}
\alias{identical.OntologySource}
\title{identical.OntologySource}
\usage{
identical.OntologySource(x, y)
}
\arguments{
\item{x}{a \link{OntologySource} object}

\item{y}{a \link{OntologySource} object}
}
\description{
Allows checking for the identity of \link{OntologySource} objects
}
