% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterValue.R
\name{identical.ParameterValue}
\alias{identical.ParameterValue}
\title{identical.ParameterValue}
\usage{
identical.ParameterValue(x, y)
}
\arguments{
\item{x}{a \code{[ParameterValue]} object}

\item{y}{a \code{[ParameterValue]} object}
}
\description{
Allows checking for the identity of \code{[ParameterValue]} objects
}
