% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\name{identical.Person}
\alias{identical.Person}
\title{identical.Person}
\usage{
identical.Person(x, y)
}
\arguments{
\item{x}{a \code{[Person]} object}

\item{y}{a \code{[Person]} object}
}
\description{
Allows checking for the identity of \code{[Person]} objects
}
