% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{identical.Process}
\alias{identical.Process}
\title{identical.Process}
\usage{
identical.Process(x, y)
}
\arguments{
\item{x}{a \code{[Process]} object}

\item{y}{a \code{[Process]} object}
}
\description{
Allows checking for the identity of \code{[Process]} objects
}
