% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protocol.R
\name{identical.Protocol}
\alias{identical.Protocol}
\title{identical.Protocol}
\usage{
identical.Protocol(x, y)
}
\arguments{
\item{x}{a \code{[Protocol]} object}

\item{y}{a \code{[Protocol]} object}
}
\description{
Allows checking for the identity of \code{[Protocol]} objects
}
