% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolComponent.R
\name{identical.ProtocolComponent}
\alias{identical.ProtocolComponent}
\title{identical.ProtocolComponent}
\usage{
identical.ProtocolComponent(x, y)
}
\arguments{
\item{x}{a \code{[ProtocolComponent]} object}

\item{y}{a \code{[ProtocolComponent]} object}
}
\description{
Allows checking for the identity of \code{[ProtocolComponent]} objects
}
