% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolParameter.R
\name{identical.ProtocolParameter}
\alias{identical.ProtocolParameter}
\title{identical.ProtocolParameter}
\usage{
identical.ProtocolParameter(x, y)
}
\arguments{
\item{x}{a \code{[ProtocolParameter]} object}

\item{y}{a \code{[ProtocolParameter]} object}
}
\description{
Allows checking for the identity of \code{[ProtocolParameter]} objects
}
