% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.R
\name{identical.Publication}
\alias{identical.Publication}
\title{identical.Publication}
\usage{
identical.Publication(x, y)
}
\arguments{
\item{x}{a \code{[Publication]} object}

\item{y}{a \code{[Publication]} object}
}
\description{
Allows checking for the identity of \code{[Publication]} objects
}
