% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{identical.Sample}
\alias{identical.Sample}
\title{identical.Sample}
\usage{
identical.Sample(x, y)
}
\arguments{
\item{x}{a \code{[Sample]} object}

\item{y}{a \code{[Sample]} object}
}
\description{
Allows checking for the identity of \code{[Sample]} objects
}
