% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{identical.Source}
\alias{identical.Source}
\title{identical.Source}
\usage{
identical.Source(x, y)
}
\arguments{
\item{x}{a \code{[Source]} object}

\item{y}{a \code{[Source]} object}
}
\description{
Allows checking for the identity of \code{[Source]} objects
}
