% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{identical.Study}
\alias{identical.Study}
\title{identical.Study}
\usage{
identical.Study(x, y)
}
\arguments{
\item{x}{a \code{[Study]} object}

\item{y}{a \code{[Study]} object}
}
\description{
Allows checking for the identity of \code{[Study]} objects
}
