% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyFactor.R
\name{identical.StudyFactor}
\alias{identical.StudyFactor}
\title{identical.StudyFactor}
\usage{
identical.StudyFactor(x, y)
}
\arguments{
\item{x}{a \code{[StudyFactor]} object}

\item{y}{a \code{[StudyFactor]} object}
}
\description{
Allows checking for the identity of \code{[StudyFactor]} objects
}
