% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit.R
\name{identical.Unit}
\alias{identical.Unit}
\title{identical.Unit}
\usage{
identical.Unit(x, y)
}
\arguments{
\item{x}{a \code{[Unit]} object}

\item{y}{a \code{[Unit]} object}
}
\description{
Allows checking for the identity of \code{[Unit]} objects
}
