% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiable.R
\name{identifiable}
\alias{identifiable}
\title{R6 superclass for identifiable objects}
\description{
option to prefix with custom name.
}
\examples{

## ------------------------------------------------
## Method `identifiable$set_id`
## ------------------------------------------------

\dontrun{
super$set_id()
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{of the id object NULL}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-identifiable-new}{\code{identifiable$new()}}
\item \href{#method-identifiable-set_id}{\code{identifiable$set_id()}}
\item \href{#method-identifiable-clone}{\code{identifiable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifiable-new"></a>}}
\if{latex}{\out{\hypertarget{method-identifiable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{identifiable$new(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{default NULL}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make a new identifiable object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifiable-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-identifiable-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{identifiable$set_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
set self id to a uuid
}

\subsection{Returns}{
null
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
super$set_id()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifiable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-identifiable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{identifiable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
