% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{index_repeated_sections}
\alias{index_repeated_sections}
\title{index_repeated_sections}
\usage{
index_repeated_sections(i_tab_with_section_type)
}
\arguments{
\item{i_tab_with_section_type}{}
}
\value{
a tibble
}
\description{
If there are multiple copies of a section headings assign them an index
number to distinguish them, this should generally only apply to studies.
}
