% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Investigation.R
\name{Investigation}
\alias{Investigation}
\title{R6 class for Investigation}
\description{
Represents an investigation from the ISA model
}
\details{
An investigation maintains metadata about the project context and links to one or more studies. There can only be 1 Investigation in an ISA descriptor.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{...}

\item{\code{identifier}}{A locally unique identifier or an accession number provided by a repository.}

\item{\code{title}}{A concise name given to the investigation.}

\item{\code{description}}{A textual description of the investigation.}

\item{\code{submission_date}}{date on which the investigation was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the investigation should be released publicly. This should be ISO8601 formatted.}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{publications}}{a list of \link{Publication} objects}

\item{\code{contacts}}{A list of People/contacts associated with an Investigation.}

\item{\code{studies}}{\link{Study} is the central unit, containing information on the subject under study.}

\item{\code{comments}}{comments associated with instances of this class.}

\item{\code{characteristic_categories}}{a \link{CharacteristicCategoryReferences} object which enumerates the possible characteristic categories}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Investigation-new}{\code{Investigation$new()}}
\item \href{#method-Investigation-check_title}{\code{Investigation$check_title()}}
\item \href{#method-Investigation-set_title}{\code{Investigation$set_title()}}
\item \href{#method-Investigation-check_description}{\code{Investigation$check_description()}}
\item \href{#method-Investigation-set_description}{\code{Investigation$set_description()}}
\item \href{#method-Investigation-check_studies}{\code{Investigation$check_studies()}}
\item \href{#method-Investigation-set_studies}{\code{Investigation$set_studies()}}
\item \href{#method-Investigation-check_contacts}{\code{Investigation$check_contacts()}}
\item \href{#method-Investigation-set_contacts}{\code{Investigation$set_contacts()}}
\item \href{#method-Investigation-check_publications}{\code{Investigation$check_publications()}}
\item \href{#method-Investigation-set_publications}{\code{Investigation$set_publications()}}
\item \href{#method-Investigation-set_public_release_date}{\code{Investigation$set_public_release_date()}}
\item \href{#method-Investigation-set_submission_date}{\code{Investigation$set_submission_date()}}
\item \href{#method-Investigation-check_ontology_source_references}{\code{Investigation$check_ontology_source_references()}}
\item \href{#method-Investigation-set_ontology_source_references}{\code{Investigation$set_ontology_source_references()}}
\item \href{#method-Investigation-check_comments}{\code{Investigation$check_comments()}}
\item \href{#method-Investigation-set_comments}{\code{Investigation$set_comments()}}
\item \href{#method-Investigation-add_comment}{\code{Investigation$add_comment()}}
\item \href{#method-Investigation-get_materials}{\code{Investigation$get_materials()}}
\item \href{#method-Investigation-get_samples}{\code{Investigation$get_samples()}}
\item \href{#method-Investigation-get_sources}{\code{Investigation$get_sources()}}
\item \href{#method-Investigation-to_table}{\code{Investigation$to_table()}}
\item \href{#method-Investigation-cat_table}{\code{Investigation$cat_table()}}
\item \href{#method-Investigation-to_list}{\code{Investigation$to_list()}}
\item \href{#method-Investigation-from_list}{\code{Investigation$from_list()}}
\item \href{#method-Investigation-from_json}{\code{Investigation$from_json()}}
\item \href{#method-Investigation-print}{\code{Investigation$print()}}
\item \href{#method-Investigation-clone}{\code{Investigation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$new(
  filename = "",
  identifier = character(),
  title = "",
  description = "",
  submission_date = NULL,
  public_release_date = NULL,
  ontology_source_references = NULL,
  publications = NULL,
  contacts = NULL,
  studies = NULL,
  comments = NULL,
  characteristic_categories = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{...}

\item{\code{identifier}}{A locally unique identifier or an accession number provided by a repository.}

\item{\code{title}}{A concise name given to the investigation.}

\item{\code{description}}{A textual description of the investigation.}

\item{\code{submission_date}}{date on which the investigation was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the investigation should be released publicly. This should be ISO8601 formatted.}

\item{\code{ontology_source_references}}{ontology_source_references \link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{publications}}{publications A list of Publications associated with an Investigation. (format?)}

\item{\code{contacts}}{contacts A list of People/contacts associated with an Investigation.}

\item{\code{studies}}{studies \link{Study} is the central unit, containing information on the subject under study.}

\item{\code{comments}}{comments comments associated with instances of this class.}

\item{\code{characteristic_categories}}{a \link{CharacteristicCategoryReferences} object which enumerates the possible characteristic categories}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new investigation object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_title"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_title}{}}}
\subsection{Method \code{check_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{The title of the investigation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the title of the investigation is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{The title of the investigation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the title of the investigation if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_description"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_description}{}}}
\subsection{Method \code{check_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the investigation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the description of the investigation is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the investigation}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the description of the investigation if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_studies"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_studies}{}}}
\subsection{Method \code{check_studies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_studies(studies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studies}}{a list of \link{Study} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check studies is a list of \link{Study} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_studies"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_studies}{}}}
\subsection{Method \code{set_studies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_studies(studies)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studies}}{a list of \link{Study} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set studies if studies is a list of \link{Study} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_contacts}{}}}
\subsection{Method \code{check_contacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_contacts(contacts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contacts}}{a list of \link{Person} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check contacts is a list of \link{Person} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_contacts}{}}}
\subsection{Method \code{set_contacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_contacts(contacts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contacts}}{a list of \link{Person} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set contacts if contacts is a list of \link{Person} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_publications"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_publications}{}}}
\subsection{Method \code{check_publications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_publications(publications)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{publications}}{a list of \link{Publication} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check publications is a list of \link{Publication} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_publications"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_publications}{}}}
\subsection{Method \code{set_publications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_publications(publications)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{publications}}{a list of \link{Publication} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check publications is a list of \link{Publication} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_public_release_date"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_public_release_date}{}}}
\subsection{Method \code{set_public_release_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_public_release_date(public_release_date, null.ok = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{public_release_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}

\item{\code{null.ok}}{accept NULL dates (boolean) default = FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set public_release_date to a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_submission_date"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_submission_date}{}}}
\subsection{Method \code{set_submission_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_submission_date(submission_date, null.ok = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}

\item{\code{null.ok}}{accept NULL dates (boolean) default = FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set submission_date to a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_ontology_source_references"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_ontology_source_references}{}}}
\subsection{Method \code{check_ontology_source_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_ontology_source_references(ontology_source_references)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that input is an \link{OntologySourceReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_ontology_source_references"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_ontology_source_references}{}}}
\subsection{Method \code{set_ontology_source_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_ontology_source_references(ontology_source_references)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set ontology_source_references to be an \link{OntologySourceReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-get_materials"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-get_materials}{}}}
\subsection{Method \code{get_materials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$get_materials()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get a list of all of the \link{Material} objects used in studies in this investigation
}

\subsection{Returns}{
list of \link{Material}s
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-get_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-get_samples}{}}}
\subsection{Method \code{get_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$get_samples()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get a list of all of the \link{Sample} objects used in studies in this investigation
}

\subsection{Returns}{
list of \link{Sample}s
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-get_sources"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-get_sources}{}}}
\subsection{Method \code{get_sources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$get_sources()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get a list of all of the \link{Source} objects used in studies in this investigation
}

\subsection{Returns}{
list of \link{Source}s
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-to_table"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-to_table}{}}}
\subsection{Method \code{to_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$to_table()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Generate a tabular representation of the investigation that can be
serialised to an ISA investigation file.
}

\subsection{Returns}{
a Tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-cat_table"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-cat_table}{}}}
\subsection{Method \code{cat_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$cat_table(path = stdout(), overwrite = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{the path/filename to which to write the output}

\item{\code{overwrite}}{should any existing files at path be overwritten?
(boolean) Default: FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
writes the tabular representation of the investigtion to a file
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Investigation} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$from_json(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path to the isa-json file}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Populate an \link{Investigation} Object from an isa-json file
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-print"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Investigation} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Investigation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Investigation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Investigation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
