% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\name{isa_person}
\alias{isa_person}
\title{R6 class for person}
\description{
Represents person
}
\section{Public fields}{

}

\examples{

## ------------------------------------------------
## Method `isa_person$to_list`
## ------------------------------------------------

person$to_list()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isa_person-new}{\code{isa_person$new()}}
\item \href{#method-isa_person-check_email}{\code{isa_person$check_email()}}
\item \href{#method-isa_person-check_orcid}{\code{isa_person$check_orcid()}}
\item \href{#method-isa_person-set_email}{\code{isa_person$set_email()}}
\item \href{#method-isa_person-set_orcid}{\code{isa_person$set_orcid()}}
\item \href{#method-isa_person-to_list}{\code{isa_person$to_list()}}
\item \href{#method-isa_person-from_list}{\code{isa_person$from_list()}}
\item \href{#method-isa_person-clone}{\code{isa_person$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-new"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$new(
  id = "",
  last_name = "",
  first_name = "",
  mid_initials = "",
  email = NULL,
  phone = "",
  fax = "",
  address = "",
  affiliation = "",
  orcid = "",
  roles = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
person
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-check_email"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-check_email}{}}}
\subsection{Method \code{check_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$check_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if email address is valid

NB the regex used to validate emails is ... difficult to represent in roxygen syntax
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-check_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-check_orcid}{}}}
\subsection{Method \code{check_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$check_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if the ORCID is valid i.e. 4 lots of 4 digits separated by hyphens
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-set_email"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-set_email}{}}}
\subsection{Method \code{set_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$set_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the email address field
email addresses ar e first checked for validity by \code{[check_email]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-set_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-set_orcid}{}}}
\subsection{Method \code{set_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$set_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the orcid field
ORCID's are first checked for validity by \code{[check_orcid]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{person$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an \code{[person]} object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[person]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isa_person-clone"></a>}}
\if{latex}{\out{\hypertarget{method-isa_person-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{isa_person$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
