% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isa_tab_section_headings}
\alias{isa_tab_section_headings}
\title{isa_tab_section_headings}
\format{
named character vector
}
\source{
https://github.com/ISA-tools/isa-specs/blob/master/source/isatab.rst
}
\usage{
isa_tab_section_headings
}
\description{
isa_tab_section_headings
}
\keyword{datasets}
