% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masters.R
\docType{data}
\name{masters}
\alias{masters}
\title{2019 Masters Tournament Scores}
\format{
A data frame with 87 rows and 9 variables:
\describe{
  \item{pos}{finishing position for a player}
  \item{player_name}{name of the golfer}
  ...
}
}
\source{
\url{https://www.augusta.com/masters/historic/leaderboards/2019}
}
\usage{
masters
}
\description{
A dataset containing the scores from the 2019 Masters Golf Tournament.
}
\keyword{datasets}
