% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masters_2025.R
\docType{data}
\name{masters_2025}
\alias{masters_2025}
\title{2025 Masters Tournament Scores}
\format{
A data frame with 95 rows and 9 variables:
\describe{
  \item{pos}{finishing position for a player}
  \item{player_name}{name of the golfer}
  \item{nationality}{nationality of the golfer}
  \item{score}{final score relative to par}
  \item{round_1}{number of strokes from the first round}
  \item{round_2}{number of strokes from the second round}
  \item{round_3}{number of strokes from the third round}
  \item{round_4}{number of strokes from the fourth round}
  \item{total}{total strokes from all rounds}
}
}
\source{
\url{https://www.espn.com/golf/leaderboard/_/tournamentId/401703504}
}
\usage{
masters_2025
}
\description{
A data frame containing the scores from the 2025 Masters Golf Tournament.
}
\keyword{datasets}
