% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Material.R
\name{Material}
\alias{Material}
\title{R6 object to represent Materials}
\description{
R6 object to represent Materials

R6 object to represent Materials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{characteristics}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Material-new}{\code{Material$new()}}
\item \href{#method-Material-check_name}{\code{Material$check_name()}}
\item \href{#method-Material-set_name}{\code{Material$set_name()}}
\item \href{#method-Material-check_characteristics}{\code{Material$check_characteristics()}}
\item \href{#method-Material-set_characteristics}{\code{Material$set_characteristics()}}
\item \href{#method-Material-get_id}{\code{Material$get_id()}}
\item \href{#method-Material-clone}{\code{Material$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-new"></a>}}
\if{latex}{\out{\hypertarget{method-Material-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$new(name = character(), type = character(), characteristics = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}

\item{\code{characteristics}}{a list of \code{[Characteristic]} objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the name of the material is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the name of the material if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_characteristics}{}}}
\subsection{Method \code{check_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \code{[Characteristic]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check characteristics is a list of \code{[Characteristic]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_characteristics}{}}}
\subsection{Method \code{set_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \code{[Characteristic]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set characteristics if characteristics is a list of \code{[Characteristic]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Material-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Material-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
