% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaterialReferences.R
\name{MaterialReferences}
\alias{MaterialReferences}
\title{R6 class for MaterialReferences}
\description{
R6 class for MaterialReferences

R6 class for MaterialReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ontology_source_references}}{...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ontology_source_references}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MaterialReferences-new}{\code{MaterialReferences$new()}}
\item \href{#method-MaterialReferences-from_list}{\code{MaterialReferences$from_list()}}
\item \href{#method-MaterialReferences-check_materials}{\code{MaterialReferences$check_materials()}}
\item \href{#method-MaterialReferences-add_materials}{\code{MaterialReferences$add_materials()}}
\item \href{#method-MaterialReferences-get_material_reference_names}{\code{MaterialReferences$get_material_reference_names()}}
\item \href{#method-MaterialReferences-clone}{\code{MaterialReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$new(material_references = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$from_list(lst, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a Material object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{MaterialReferences} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-check_materials"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-check_materials}{}}}
\subsection{Method \code{check_materials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$check_materials(materials)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Material}}{a list of \link{Material} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check Material is a list of \link{Material} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-add_materials"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-add_materials}{}}}
\subsection{Method \code{add_materials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$add_materials(materials)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Add ontology sources to the
}

\subsection{Returns}{
a vector of ontology source names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-get_material_reference_names"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-get_material_reference_names}{}}}
\subsection{Method \code{get_material_reference_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$get_material_reference_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the names of the Ontology Sources Used in this Investigation
}

\subsection{Returns}{
a vector of ontology source names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MaterialReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MaterialReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MaterialReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
