% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_four_factors_team_2021.R
\docType{data}
\name{nba_ff_team_2021}
\alias{nba_ff_team_2021}
\title{2021 NBA Four Factors (Team) for the 2020-21 Season}
\format{
A data frame with 2160 rows and 11 variables:
\describe{
  \item{idGame}{Game ID on nba.com}
  \item{slugTeam}{Team abbreviation}
  \item{idTeam}{Team ID on nba.com}
  \item{pctEFGOpponent}{Effective field goal percentage of opponent}
  \item{pctTOVOpponent}{Turnover percentage of opponent}
  \item{pctOREBOpponent}{Offensive rebound percentage of opponent}
  \item{teamName}{Team Mascot}
  \item{cityTeam}{City of slugTeam}
  \item{minExact}{Total minutes played}
  \item{rateFTA}{Free throw rate of slugTeam}
  \item{rateFTAOpponent}{Free throw rate of opponent}
}
}
\source{
\url{http://asbcllc.com/nbastatR/reference/box_scores.html}
}
\usage{
nba_ff_team_2021
}
\description{
A dataset containing the NBA Four Factor table on nba.com
}
\keyword{datasets}
