% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwsl_player_stats_2023.R
\docType{data}
\name{nwsl_player_stats_2023}
\alias{nwsl_player_stats_2023}
\title{2023 NWSL Player Statistics}
\format{
A data frame with 307 rows and 18 variables:
\describe{
  \item{player}{Player name}
  \item{nation}{Player home country}
  \item{pos}{Player position}
  \item{squad}{Player team}
  \item{age}{Age of player}
  \item{mp}{Matches played}
  \item{starts}{Number of matches in which player started the game}
  \item{min}{Total minutes played in the season}
  \item{Gp90}{Goals per ninety minutes}
  \item{Ap90}{Assists per ninety minutes}
  \item{GAp90}{Goals plus assists per ninety minutes}
  \item{npGp90}{Goals minus penalty goals per ninety minutes}
  \item{npGAp90}{Goals plus assists minus penalty goals per ninety minutes}
  \item{xGp90}{Expected goals per ninety minutes}
  \item{xAp90}{Expected assists per ninety minutes}
  \item{xGxAp90}{Expected goals plus assists per ninety minutes}
  \item{npxGp90}{Expected goals minus penalty goals per ninety minutes}
  \item{npxGxAp90}{Expected goals plus assists minus penalty goals per ninety minutes}
}
}
\source{
\url{https://fbref.com/en/comps/182/stats/NWSL-Stats}
}
\usage{
nwsl_player_stats_2023
}
\description{
A dataset containing individual player statistics for the National Women's Soccer League
}
\keyword{datasets}
