% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OM}
\alias{OM}
\title{Ontology of Units of Measure}
\format{
\code{[OntologySource]} object
}
\source{
\url{https://raw.githubusercontent.com/HajoRijgersberg/OM/master/om-2.0.rdf}
}
\usage{
OM
}
\description{
Ontology of Units of Measure
}
\keyword{datasets}
