% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{ontology_annotation_with_undefined_source_handler}
\alias{ontology_annotation_with_undefined_source_handler}
\title{ontology_annotation_with_undefined_source_handler}
\usage{
ontology_annotation_with_undefined_source_handler(
  self,
  lst,
  member,
  ontology_source_name = "UnknownOntologySource",
  source_description
)
}
\arguments{
\item{self}{object}

\item{lst}{ontology anotation components list}

\item{member}{the member of the object being evaluated}

\item{ontology_source_name}{the name of the auto-generated OntologySource}

\item{source_description}{the description of the auto-generated OntologySource}
}
\description{
ontology_annotation_with_undefined_source_handler
}
