% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySource.R
\name{OntologySource}
\alias{OntologySource}
\title{R6 class for the source of ontology terms}
\description{
An \link{OntologySource} describes the resource from which the value of an \link{OntologyAnnotation} is derived from.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the source of a term; i.e. the source controlled vocabulary or ontology.}

\item{\code{file}}{A file name or a URI of an official resource.}

\item{\code{file_type}}{the file type, parsable by \code{rdflib::rdf_parse()}}

\item{\code{url}}{...}

\item{\code{version}}{version of the ontology}

\item{\code{description}}{free text description of this source of ontology terms}

\item{\code{get_terms_list}}{function to generate valid terms for file}

\item{\code{terms_list}}{list of valid terms where keys are terms and values accessions}

\item{\code{comments}}{comments}

\item{\code{explicitly_provided}}{Was this \link{OntologySource} provided in the
list of \link{OntologySourceReferences} for the \link{Investigation} (TRUE) or was
it dynamically generated (FALSE)?}

\item{\code{origin}}{the origin of the \link{OntologySource} object, was it listed in
an investigation ontolgoy origin reference list, elsewhere or auto generated}

\item{\code{isa_process_type}}{the type of process from the list of predefined
types see: \code{names(valid_isa_process_types)}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OntologySource-new}{\code{OntologySource$new()}}
\item \href{#method-OntologySource-check_name}{\code{OntologySource$check_name()}}
\item \href{#method-OntologySource-set_name}{\code{OntologySource$set_name()}}
\item \href{#method-OntologySource-check_file}{\code{OntologySource$check_file()}}
\item \href{#method-OntologySource-set_file}{\code{OntologySource$set_file()}}
\item \href{#method-OntologySource-check_file_type}{\code{OntologySource$check_file_type()}}
\item \href{#method-OntologySource-set_file_type}{\code{OntologySource$set_file_type()}}
\item \href{#method-OntologySource-check_url}{\code{OntologySource$check_url()}}
\item \href{#method-OntologySource-set_url}{\code{OntologySource$set_url()}}
\item \href{#method-OntologySource-check_version}{\code{OntologySource$check_version()}}
\item \href{#method-OntologySource-set_version}{\code{OntologySource$set_version()}}
\item \href{#method-OntologySource-check_description}{\code{OntologySource$check_description()}}
\item \href{#method-OntologySource-set_description}{\code{OntologySource$set_description()}}
\item \href{#method-OntologySource-check_comments}{\code{OntologySource$check_comments()}}
\item \href{#method-OntologySource-set_comments}{\code{OntologySource$set_comments()}}
\item \href{#method-OntologySource-add_comment}{\code{OntologySource$add_comment()}}
\item \href{#method-OntologySource-add_terms}{\code{OntologySource$add_terms()}}
\item \href{#method-OntologySource-to_list}{\code{OntologySource$to_list()}}
\item \href{#method-OntologySource-set_isa_process_type}{\code{OntologySource$set_isa_process_type()}}
\item \href{#method-OntologySource-from_list}{\code{OntologySource$from_list()}}
\item \href{#method-OntologySource-is_same_ontology_as}{\code{OntologySource$is_same_ontology_as()}}
\item \href{#method-OntologySource-print}{\code{OntologySource$print()}}
\item \href{#method-OntologySource-clone}{\code{OntologySource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-new"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$new(
  name = character(),
  file = character(),
  file_type = character(),
  url = character(),
  version = character(),
  description = character(),
  get_terms_list = NULL,
  terms_list = NULL,
  comments = NULL,
  explicitly_provided = logical(),
  origin = character(),
  isa_process_type = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the source of a term; i.e. the source controlled vocabulary or ontology.}

\item{\code{file}}{A file name or a URI of an official resource.}

\item{\code{file_type}}{the file type, parsable by \code{rdflib::rdf_parse()}}

\item{\code{url}}{...}

\item{\code{version}}{vesion of the ontology}

\item{\code{description}}{free text description of this source of ontology terms}

\item{\code{get_terms_list}}{function to generate valid terms for file}

\item{\code{terms_list}}{list of valid terms where keys are terms and values accessions}

\item{\code{comments}}{comments}

\item{\code{explicitly_provided}}{Was this \link{OntologySource} provided in the
list of \link{OntologySourceReferences} for the \link{Investigation} (TRUE) or was
it dynamically generated (FALSE)?}

\item{\code{origin}}{the source of the \link{OntologySource} object, was it listed in
an investigation ontolgoy source reference list, elsewhere or auto generated
an investigation ontolgoy source reference list, elsewhere or auto generated}

\item{\code{isa_process_type}}{the type of process from the list of predefined}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{OntologySource} object to act as source for ontology terms
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the name of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the name of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_file"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_file}{}}}
\subsection{Method \code{check_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the file of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_file"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_file}{}}}
\subsection{Method \code{set_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{The file of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the file of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_file_type"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_file_type}{}}}
\subsection{Method \code{check_file_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_file_type(file_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_type}}{The file_type of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the file_type of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_file_type"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_file_type}{}}}
\subsection{Method \code{set_file_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_file_type(file_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_type}}{The file_type of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the file_type of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_url"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_url}{}}}
\subsection{Method \code{check_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_url(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The url of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the url of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_url"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_url}{}}}
\subsection{Method \code{set_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_url(url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The url of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the url of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_version"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_version}{}}}
\subsection{Method \code{check_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{The version of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the version of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_version"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_version}{}}}
\subsection{Method \code{set_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{The version of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the version of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_description"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_description}{}}}
\subsection{Method \code{check_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the description of the \link{OntologySource} is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the description of the \link{OntologySource} if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-add_terms"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-add_terms}{}}}
\subsection{Method \code{add_terms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$add_terms(terms_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{terms_list}}{a list of terms with terms accessions as values}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add terms to the \link{OntologySource}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$to_list(json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{json}}{(default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-set_isa_process_type"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-set_isa_process_type}{}}}
\subsection{Method \code{set_isa_process_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$set_isa_process_type(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the type name}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the type of the process from the list of valid pre-defined
types. This is used to given certain types of process custom
names in the isa-tab serialisation.
see \code{names(valid_isa_process_types)} for valid types
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$from_list(lst, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}

\item{\code{json}}{(default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{OntologySource} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-is_same_ontology_as"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-is_same_ontology_as}{}}}
\subsection{Method \code{is_same_ontology_as()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$is_same_ontology_as(other_ontology_source)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other_ontology_source}}{another \link{OntologySource}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
test if \link{OntologySource} is equivalent another OntologySource
not necessarily an exactly identical object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-print"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints an \link{OntologySource} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
