% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ow_golf_rankings.R
\docType{data}
\name{ow_golf_rankings}
\alias{ow_golf_rankings}
\title{2022 Official World Golf Rankings}
\format{
A data frame with 300 rows and 2 variables:
\describe{
  \item{WGR_June_2022}{official golf world ranking}
  \item{Player_initial_last}{initial of first name and complete last name of player}
}
}
\source{
\url{https://www.owgr.com/archive/archive.sps?iType=1747}
}
\usage{
ow_golf_rankings
}
\description{
A dataset containing the Official World Golf Rankings in June 2022
}
\keyword{datasets}
