% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{parse_investigation}
\alias{parse_investigation}
\title{parse_investigation}
\usage{
parse_investigation(path, section_headings = isa_tab_section_headings)
}
\arguments{
\item{path}{the path to an investigation file}

\item{section_headings}{named character vector of section headings}
}
\value{
a nested tibble
}
\description{
parse_investigation
}
