% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\name{Person}
\alias{Person}
\title{R6 class for Person}
\description{
Represents a person
}
\examples{

## ------------------------------------------------
## Method `Person$to_list`
## ------------------------------------------------

Person$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Person-new}{\code{Person$new()}}
\item \href{#method-Person-check_email}{\code{Person$check_email()}}
\item \href{#method-Person-check_orcid}{\code{Person$check_orcid()}}
\item \href{#method-Person-set_email}{\code{Person$set_email()}}
\item \href{#method-Person-set_orcid}{\code{Person$set_orcid()}}
\item \href{#method-Person-to_list}{\code{Person$to_list()}}
\item \href{#method-Person-from_list}{\code{Person$from_list()}}
\item \href{#method-Person-clone}{\code{Person$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-new"></a>}}
\if{latex}{\out{\hypertarget{method-Person-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$new(
  id = "",
  last_name = "",
  first_name = "",
  mid_initials = "",
  email = "",
  phone = "",
  fax = "",
  address = "",
  affiliation = "",
  orcid = "",
  roles = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
person
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-check_email"></a>}}
\if{latex}{\out{\hypertarget{method-Person-check_email}{}}}
\subsection{Method \code{check_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$check_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if email address is valid

NB the regex used to validate emails is ... difficult to represent in roxygen syntax
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-check_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-Person-check_orcid}{}}}
\subsection{Method \code{check_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$check_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if the ORCID is valid i.e. 4 lots of 4 digits separated by hyphens
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_email"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_email}{}}}
\subsection{Method \code{set_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the email address field
email addresses ar e first checked for validity by \code{[check_email]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_orcid}{}}}
\subsection{Method \code{set_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the orcid field
ORCID's are first checked for validity by \code{[check_orcid]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Person-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Person$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Person-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an \code{[Person]} object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Person]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Person-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
