% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pretty_print_comments}
\alias{pretty_print_comments}
\title{pretty_print_comments}
\usage{
pretty_print_comments(comments)
}
\arguments{
\item{comments}{a named list}
}
\description{
Green bold section heading
indented comments with bold titles and wrapped contents
}
