% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{process_io_paths}
\alias{process_io_paths}
\title{process_io_paths}
\usage{
process_io_paths(processes)
}
\arguments{
\item{processes}{list of \link{Process} objects as in the process_sequence
sequence field of an \link{Assay} object}
}
\description{
process_io_paths
}
