% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{process_order}
\alias{process_order}
\title{process_order}
\usage{
process_order(processes)
}
\arguments{
\item{processes}{list of \link{Process} objects as in the process_sequence
sequence field of an \link{Assay} object}
}
\description{
process_order
}
