% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSequenceNode.R
\name{ProcessSequenceNode}
\alias{ProcessSequenceNode}
\title{R6 class for ProcessSequenceNode}
\description{
R6 class for ProcessSequenceNode

R6 class for ProcessSequenceNode
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProcessSequenceNode-new}{\code{ProcessSequenceNode$new()}}
\item \href{#method-ProcessSequenceNode-assign_identifier}{\code{ProcessSequenceNode$assign_identifier()}}
\item \href{#method-ProcessSequenceNode-get_identifier}{\code{ProcessSequenceNode$get_identifier()}}
\item \href{#method-ProcessSequenceNode-clone}{\code{ProcessSequenceNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessSequenceNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessSequenceNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessSequenceNode$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
new ProcessSequenceNode
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessSequenceNode-assign_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessSequenceNode-assign_identifier}{}}}
\subsection{Method \code{assign_identifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessSequenceNode$assign_identifier()}\if{html}{\out{</div>}}
}

\subsection{Details}{
increment ProcessSequenceNode
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessSequenceNode-get_identifier"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessSequenceNode-get_identifier}{}}}
\subsection{Method \code{get_identifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessSequenceNode$get_identifier()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get ProcessSequenceNode identifier
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProcessSequenceNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProcessSequenceNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProcessSequenceNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
