% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolReferences.R
\name{ProtocolReferences}
\alias{ProtocolReferences}
\title{ProtocolReferences}
\description{
ProtocolReferences

ProtocolReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{protocols}}{A names list of \link{Protocol} objects}

\item{\code{ontology_source_references}}{ontology_source_references}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProtocolReferences-new}{\code{ProtocolReferences$new()}}
\item \href{#method-ProtocolReferences-check_protocols}{\code{ProtocolReferences$check_protocols()}}
\item \href{#method-ProtocolReferences-add_protocols}{\code{ProtocolReferences$add_protocols()}}
\item \href{#method-ProtocolReferences-set_protocols}{\code{ProtocolReferences$set_protocols()}}
\item \href{#method-ProtocolReferences-get_protocol_ids}{\code{ProtocolReferences$get_protocol_ids()}}
\item \href{#method-ProtocolReferences-get_protocol_names}{\code{ProtocolReferences$get_protocol_names()}}
\item \href{#method-ProtocolReferences-get_protocol_origins}{\code{ProtocolReferences$get_protocol_origins()}}
\item \href{#method-ProtocolReferences-n_protocols}{\code{ProtocolReferences$n_protocols()}}
\item \href{#method-ProtocolReferences-to_list}{\code{ProtocolReferences$to_list()}}
\item \href{#method-ProtocolReferences-from_list}{\code{ProtocolReferences$from_list()}}
\item \href{#method-ProtocolReferences-print}{\code{ProtocolReferences$print()}}
\item \href{#method-ProtocolReferences-clone}{\code{ProtocolReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$new(protocols = list(), ontology_source_references = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocols}}{A names list of \link{Protocol} objects}

\item{\code{ontology_source_references}}{ontology_source_references}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{ProtocolReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-check_protocols"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-check_protocols}{}}}
\subsection{Method \code{check_protocols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$check_protocols(protocols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocols}}{a list of \link{Protocol} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check protocols is a list of \link{Protocols} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-add_protocols"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-add_protocols}{}}}
\subsection{Method \code{add_protocols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$add_protocols(protocols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocols}}{a list of \link{Protocols} to add to the
\link{ProtocolReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add \link{Protocols} to the \link{ProtocolsReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-set_protocols"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-set_protocols}{}}}
\subsection{Method \code{set_protocols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$set_protocols(protocols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocols}}{a list of \link{Protocols} to add to the
\link{ProtocolReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the \link{Protocols} to the \link{ProtocolsReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-get_protocol_ids"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-get_protocol_ids}{}}}
\subsection{Method \code{get_protocol_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$get_protocol_ids()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the IDs of the protocols in the reference
}

\subsection{Returns}{
a character vector of protocol IDs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-get_protocol_names"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-get_protocol_names}{}}}
\subsection{Method \code{get_protocol_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$get_protocol_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the names of the protocols in the reference
}

\subsection{Returns}{
a character vector of protocol namees
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-get_protocol_origins"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-get_protocol_origins}{}}}
\subsection{Method \code{get_protocol_origins()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$get_protocol_origins()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the origins of the protocols in the reference
}

\subsection{Returns}{
a character vector of the origins of the protocols in the
reference
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-n_protocols"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-n_protocols}{}}}
\subsection{Method \code{n_protocols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$n_protocols()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the number of protocols in the reference
}

\subsection{Returns}{
an integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
serialise the ProtocolReferences object to a list
}

\subsection{Returns}{
a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$from_list(lst, origin = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list}

\item{\code{origin}}{the id of the}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Populate the fields of the \link{ProtocolReferences} object using a list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-print"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints the \link{ProtocolReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProtocolReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProtocolReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProtocolReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
