% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.R
\name{Publication}
\alias{Publication}
\title{R6 class for a Publication}
\description{
A publication associated with an investigation or study.
}
\details{
possible use of  https://docs.ropensci.org/rcrossref/index.html

consider a strict mode where the ID must be able to be identified in suitable databases

Create a new [Publication Object

Check for a valid PubMed ID

Set PubMed ID if valid

Check for a valid DOI (Digital Object Identifier)

Set the DOI (Digital Object Identifier) if valid

Check that all Authors listed are represented with [Person objects

Set the list of authors of the Investigation or Study if they are valid

Check the the title has a non-zero length

Set the title of the [Publication

checks if comments are a named list of character vectors

Sets comments if they are in a valid format

Add comment if it is in a valid format

convert the Publication object to a list

Generate a [Publication object from a list

Pretty prints \link{Publication} objects
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pubmed_id}}{The PubMed IDs of the described publication(s) associated with this investigation.}

\item{\code{doi}}{A Digital Object Identifier (DOI) for that publication (where available).}

\item{\code{author_list}}{The list of authors associated with that publication.}

\item{\code{title}}{The title of publication associated with the investigation.}

\item{\code{status}}{A term describing the status of that publication (i.e.submitted, in preparation, published).}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Publication-new}{\code{Publication$new()}}
\item \href{#method-Publication-check_pubmed_id}{\code{Publication$check_pubmed_id()}}
\item \href{#method-Publication-set_pubmed_id}{\code{Publication$set_pubmed_id()}}
\item \href{#method-Publication-check_doi}{\code{Publication$check_doi()}}
\item \href{#method-Publication-set_doi}{\code{Publication$set_doi()}}
\item \href{#method-Publication-check_author_list}{\code{Publication$check_author_list()}}
\item \href{#method-Publication-set_author_list}{\code{Publication$set_author_list()}}
\item \href{#method-Publication-check_title}{\code{Publication$check_title()}}
\item \href{#method-Publication-set_title}{\code{Publication$set_title()}}
\item \href{#method-Publication-check_comments}{\code{Publication$check_comments()}}
\item \href{#method-Publication-set_comments}{\code{Publication$set_comments()}}
\item \href{#method-Publication-add_comment}{\code{Publication$add_comment()}}
\item \href{#method-Publication-to_list}{\code{Publication$to_list()}}
\item \href{#method-Publication-from_list}{\code{Publication$from_list()}}
\item \href{#method-Publication-print}{\code{Publication$print()}}
\item \href{#method-Publication-clone}{\code{Publication$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-new"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$new(
  pubmed_id = NULL,
  doi = character(),
  author_list = NULL,
  title = character(),
  status = NULL,
  comments = NULL,
  ontology_source_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pubmed_id}}{pubmed_id The PubMed IDs of the described publication(s) associated with this investigation.}

\item{\code{doi}}{doi A Digital Object Identifier (DOI) for that publication (where available).}

\item{\code{author_list}}{author_list The list of authors associated with that publication.}

\item{\code{title}}{title The title of publication associated with the investigation.}

\item{\code{status}}{status A term describing the status of that publication (i.e.submitted, in preparation, published).}

\item{\code{comments}}{comments Comments associated with instances of this class.}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-check_pubmed_id"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-check_pubmed_id}{}}}
\subsection{Method \code{check_pubmed_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$check_pubmed_id(pubmed_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pubmed_id}}{an identifier from the PubMed database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-set_pubmed_id"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-set_pubmed_id}{}}}
\subsection{Method \code{set_pubmed_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$set_pubmed_id(pubmed_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pubmed_id}}{an identifier from the PubMed database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-check_doi"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-check_doi}{}}}
\subsection{Method \code{check_doi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$check_doi(doi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doi}}{a valid DOI (Digital Object Identifier)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-set_doi"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-set_doi}{}}}
\subsection{Method \code{set_doi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$set_doi(doi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{doi}}{a valid DOI (Digital Object Identifier)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-check_author_list"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-check_author_list}{}}}
\subsection{Method \code{check_author_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$check_author_list(author_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{author_list}}{A list of authors of the Investigation or Study (list of Person objects)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-set_author_list"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-set_author_list}{}}}
\subsection{Method \code{set_author_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$set_author_list(author_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{author_list}}{a list of authors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-check_title"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-check_title}{}}}
\subsection{Method \code{check_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$check_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{of the publication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$set_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{of the publication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list suitable for conversion to a Publication Object}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-print"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Publication-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Publication-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Publication$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
