% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{read_investigation_and_assign_lines_to_sections}
\alias{read_investigation_and_assign_lines_to_sections}
\title{read_investigation_and_assign_lines_to_sections}
\usage{
read_investigation_and_assign_lines_to_sections(
  path,
  section_headings = isa_tab_section_headings
)
}
\arguments{
\item{path}{path to investigation file}

\item{section_headings}{named character vector of section headings}
}
\value{
a tibble
}
\description{
read_investigation_and_assign_lines_to_sections
}
