% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{recursive_sort_list_by_name}
\alias{recursive_sort_list_by_name}
\title{recursive_sort_list_by_name}
\usage{
recursive_sort_list_by_name(x)
}
\arguments{
\item{x}{a list}
}
\description{
recursive_sort_list_by_name
}
