% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{s3_identical_maker}
\alias{s3_identical_maker}
\title{s3_identical_maker}
\usage{
s3_identical_maker(obj_pub_props, get_id = TRUE)
}
\arguments{
\item{obj_pub_props}{a chacater vector of public properties whose values
should be compared to determine identity}

\item{get_id}{(default = TRUE) Compare the object's unique identifiers
using the get_id method}
}
\description{
function to aid in the creation of S3 generic methods for \code{[identical]}
or R6 objects
}
