% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySourceReferences.R
\name{set_ontology_source_references}
\alias{set_ontology_source_references}
\title{set_ontology_source_references}
\usage{
set_ontology_source_references(
  self,
  ontology_source_references,
  null.action = "error"
)
}
\arguments{
\item{self}{an object with an ontology_source_references attribute}

\item{ontology_source_references}{an \link{OntologySourceReferences} object}

\item{null.action}{how to handle NULLs:
\itemize{
\item "error" thow an error
\item "passthrough" set to NULL
\item "create" set to an empty  \link{OntologySourceReferences} object
}}
}
\description{
sets ontology_source_references attribute if
ontology_source_reference is an \link{OntologySourceReferences} object
}
