% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtocolReferences.R
\name{set_protocol_references}
\alias{set_protocol_references}
\title{set_protocol_references}
\usage{
set_protocol_references(self, protocol_references, null.action = "error")
}
\arguments{
\item{self}{an object with an protocol_references attribute}

\item{protocol_references}{an \link{ProtocolReferences} object}

\item{null.action}{how to handle NULLs:
\itemize{
\item "error" thow an error
\item "passthrough" set to NULL
\item "create" set to an empty  \link{ProtocolReferences} object
}}
}
\description{
sets protocol_references attribute if
protocol_reference is an \link{ProtocolReferences} object
}
