% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitReferences.R
\name{set_unit_references}
\alias{set_unit_references}
\title{set_unit_references}
\usage{
set_unit_references(self, unit_references, null.action = "error")
}
\arguments{
\item{self}{an object with an unit_references attribute}

\item{unit_references}{an \link{UnitReferences} object}

\item{null.action}{how to handle NULLs:
\itemize{
\item "error" thow an error
\item "passthrough" set to NULL
\item "create" set to an empty  \link{UnitReferences} object
}}
}
\description{
sets unit_references attribute if
unit_reference is an \link{UnitReferences} object
}
