% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{source}
\alias{source}
\title{R6 class for a source}
\description{
R6 class for a source

R6 class for a source
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-source-new}{\code{source$new()}}
\item \href{#method-source-to_list}{\code{source$to_list()}}
\item \href{#method-source-from_list}{\code{source$from_list()}}
\item \href{#method-source-clone}{\code{source$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source-new"></a>}}
\if{latex}{\out{\hypertarget{method-source-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{source$new(id = "", name = "", characteristics = list(), comments = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-source-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{source$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)
#' @details

Make \code{[source]} from list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-source-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{source$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a source object serialized to a list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-source-clone"></a>}}
\if{latex}{\out{\hypertarget{method-source-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{source$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
