% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{split_study_or_assay_by_object}
\alias{split_study_or_assay_by_object}
\title{split_study_or_assay_by_object}
\usage{
split_study_or_assay_by_object(data)
}
\arguments{
\item{data}{a}
}
\value{
a list of tibbles
}
\description{
Objects in Study and Assay tables are 'generally'* defined by a
leading column of the form: '<Entity> Name'
}
\details{
!! problem for reliablly identifying object boundaries in these tables
!! is there a more systematic way which does not make the name assumption?
!! seems like it would be complicated and by exclusion rather than a positive  ID?
}
