% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_factor.R
\name{study_factor}
\alias{study_factor}
\title{R6 class for an experimental factor}
\description{
A \code{study_factor} corresponds to an independent variable manipulated by the
experimentalist with the intention to affect biological systems in a way
that can be measured by an assay.
}
\section{Public fields}{

}

\examples{

## ------------------------------------------------
## Method `study_factor$to_list`
## ------------------------------------------------

study_factor$to_list()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[ontology_source]} reference of the study factor type}

\item{\code{id}}{unique identifier...}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-study_factor-new}{\code{study_factor$new()}}
\item \href{#method-study_factor-to_list}{\code{study_factor$to_list()}}
\item \href{#method-study_factor-from_list}{\code{study_factor$from_list()}}
\item \href{#method-study_factor-clone}{\code{study_factor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-new"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$new(
  name = "",
  factor_type = ontology_annotation$new(),
  id = "",
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[ontology_source]} reference of the study factor_type}

\item{\code{id}}{unique identifier...}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new study factor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{study_factor$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a study factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{ontology_annotation} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
