% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyFactorReferences.R
\name{StudyFactorReferences}
\alias{StudyFactorReferences}
\title{R6 Class for StudyFactorReferences}
\description{
R6 Class for StudyFactorReferences

R6 Class for StudyFactorReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{study_factor_references}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyFactorReferences-new}{\code{StudyFactorReferences$new()}}
\item \href{#method-StudyFactorReferences-check_study_factors}{\code{StudyFactorReferences$check_study_factors()}}
\item \href{#method-StudyFactorReferences-add_study_factors}{\code{StudyFactorReferences$add_study_factors()}}
\item \href{#method-StudyFactorReferences-to_list}{\code{StudyFactorReferences$to_list()}}
\item \href{#method-StudyFactorReferences-from_list}{\code{StudyFactorReferences$from_list()}}
\item \href{#method-StudyFactorReferences-get_study_factor_names}{\code{StudyFactorReferences$get_study_factor_names()}}
\item \href{#method-StudyFactorReferences-get_study_factor_ids}{\code{StudyFactorReferences$get_study_factor_ids()}}
\item \href{#method-StudyFactorReferences-print}{\code{StudyFactorReferences$print()}}
\item \href{#method-StudyFactorReferences-clone}{\code{StudyFactorReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$new(
  study_factor_references = NULL,
  ontology_source_references = NULL,
  unit_references = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-check_study_factors"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-check_study_factors}{}}}
\subsection{Method \code{check_study_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$check_study_factors(study_factors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-add_study_factors"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-add_study_factors}{}}}
\subsection{Method \code{add_study_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$add_study_factors(study_factors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$from_list(lst, explicitly_provided = logical())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-get_study_factor_names"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-get_study_factor_names}{}}}
\subsection{Method \code{get_study_factor_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$get_study_factor_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-get_study_factor_ids"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-get_study_factor_ids}{}}}
\subsection{Method \code{get_study_factor_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$get_study_factor_ids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-print"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
