% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supervision}
\alias{supervision}
\title{CRediT taxonomy: Supervision}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/supervision/}
}
\usage{
supervision
}
\description{
CRediT taxonomy: Supervision
}
\keyword{datasets}
