% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing-utility.R
\name{test_comments}
\alias{test_comments}
\title{test_comments}
\usage{
test_comments(obj)
}
\arguments{
\item{obj}{an object with comments}
}
\description{
Perform multiple standard tests on objects which use the comment methods:
\code{check_comments}, \code{set_comments}, & \code{add_comment}
}
