% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{test_empty}
\alias{test_empty}
\title{test_empty}
\usage{
test_empty(x, mode, null.ok = FALSE)
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode}

\item{null.ok}{allow NULL to be valid}
}
\value{
TRUE or FALSE
}
\description{
test_empty
}
\examples{

mode <- "character"
test_empty(character(), mode)
test_empty(character(), mode, null.ok = TRUE)
test_empty(NULL, mode, null.ok = TRUE)
test_empty("", mode)
test_empty("", mode, null.ok = TRUE)

}
