% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{test_id}
\alias{test_id}
\title{test_id
Tests if an id is valid, returns TRUE if valid and FALSE if not
(follows checkmate conventions)}
\usage{
test_id(id)
}
\arguments{
\item{id}{a uuid or uuid with a suffix}
}
\description{
test_id
Tests if an id is valid, returns TRUE if valid and FALSE if not
(follows checkmate conventions)
}
