% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process.R
\name{to_table_by_process_io_type}
\alias{to_table_by_process_io_type}
\title{to_table_by_process_io_type}
\usage{
to_table_by_process_io_type(x)
}
\arguments{
\item{x}{a list of objects which can be \link{Process}es or their inputs and outputs}
}
\description{
generate tabular outputs for objects of type: \link{DateFile}, \link{Sample},
\link{Process}, \link{Source}, or \link{Material}
}
