% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajson-parser.R
\name{unique_role_sources}
\alias{unique_role_sources}
\title{unique_role_sources}
\usage{
unique_role_sources(roles)
}
\arguments{
\item{roles}{list of roles}
}
\value{
vector of unique role sources
}
\description{
unique_role_sources
}
