% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitReferences.R
\name{UnitReferences}
\alias{UnitReferences}
\title{R6 class for UnitReferences}
\description{
R6 class for UnitReferences

R6 class for UnitReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{units}}{A list of units used as a \link{UnitReferences} object}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnitReferences-new}{\code{UnitReferences$new()}}
\item \href{#method-UnitReferences-check_unit_references}{\code{UnitReferences$check_unit_references()}}
\item \href{#method-UnitReferences-set_unit_references}{\code{UnitReferences$set_unit_references()}}
\item \href{#method-UnitReferences-add_unit_references}{\code{UnitReferences$add_unit_references()}}
\item \href{#method-UnitReferences-get_unit_ids}{\code{UnitReferences$get_unit_ids()}}
\item \href{#method-UnitReferences-get_unit_types}{\code{UnitReferences$get_unit_types()}}
\item \href{#method-UnitReferences-get_unit_origins}{\code{UnitReferences$get_unit_origins()}}
\item \href{#method-UnitReferences-to_list}{\code{UnitReferences$to_list()}}
\item \href{#method-UnitReferences-from_list}{\code{UnitReferences$from_list()}}
\item \href{#method-UnitReferences-print}{\code{UnitReferences$print()}}
\item \href{#method-UnitReferences-clone}{\code{UnitReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$new(units = NULL, ontology_source_references = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{A list of units used as a \link{Unit} object}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-check_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-check_unit_references}{}}}
\subsection{Method \code{check_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$check_unit_references(units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{a list of \link{Unit} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if this input is a list of \link{Unit} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-set_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-set_unit_references}{}}}
\subsection{Method \code{set_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$set_unit_references(units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{a list of \link{Unit} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set new units of the unit reference, overwrites current units
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-add_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-add_unit_references}{}}}
\subsection{Method \code{add_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$add_unit_references(units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{a list of \link{Unit} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add new units to the unit reference
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-get_unit_ids"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-get_unit_ids}{}}}
\subsection{Method \code{get_unit_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$get_unit_ids(origin = "any")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{origin}}{the ids of the origin of the units for which you would like to get ids}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the @ids of the unit references
}

\subsection{Returns}{
character vector of unit reference @ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-get_unit_types"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-get_unit_types}{}}}
\subsection{Method \code{get_unit_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$get_unit_types(origin = "any")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{origin}}{the ids of the origin of the units for which you would like to get types}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Get the types of the unit references
}

\subsection{Returns}{
character vector of unit types
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-get_unit_origins"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-get_unit_origins}{}}}
\subsection{Method \code{get_unit_origins()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$get_unit_origins()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the origin of the unit e.g. from which study's list of unit
categories did this reference originate
}

\subsection{Returns}{
character vector of unit origins
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$to_list(origin = "any")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{origin}}{ids of origins of unit references for which
to generate a list represenation of unit references.
"any" lists categories from all origins.
default = "any"}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Generate an R list representation of a \link{UnitReferences} object
}

\subsection{Returns}{
An R list representation of a \link{UnitReferences} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$from_list(lst, origin = NA, add = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a \link{UnitReferences} object serialized to a list}

\item{\code{origin}}{the origin of the \link{Unit} object, if it was it listed
in study unit categories list which one?}

\item{\code{add}}{(logical) add new ob}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make a \link{UnitReferences} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-print"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{UnitReferences} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
