% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visualization}
\alias{visualization}
\title{CRediT taxonomy: Visualization}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/visualization/}
}
\usage{
visualization
}
\description{
CRediT taxonomy: Visualization
}
\keyword{datasets}
