% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_player_shots_2024.R
\docType{data}
\name{wnba_player_shots_2024}
\alias{wnba_player_shots_2024}
\title{2024 WNBA Player Shots}
\format{
A data frame with 3388 rows and 3 variables:
\describe{
  \item{athlete_display_name}{Player name}
.  \item{coordinate_x}{X Coordinate of Shot}
.  \item{coordinate_y}{Y Coordinate of Shot}
}
}
\source{
\url{https://wehoop.sportsdataverse.org}
}
\usage{
wnba_player_shots_2024
}
\description{
A dataset containing Breanna Stewart and A'ja Wilson's WNBA shots in the 2024 season
}
\keyword{datasets}
