% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wnba_seattle_pbp_2024.R
\docType{data}
\name{wnba_seattle_pbp_2024}
\alias{wnba_seattle_pbp_2024}
\title{2024 WNBA Seattle Storm Shots}
\format{
A data frame with 5156 rows and 3 variables:
\describe{
.  \item{coordinate_x}{X Coordinate of Shot}
.  \item{coordinate_y}{Y Coordinate of Shot}
  \item{scoring_play}{Made or Missed Shot}
}
}
\source{
\url{https://wehoop.sportsdataverse.org}
}
\usage{
wnba_seattle_pbp_2024
}
\description{
A dataset containing all Seattle Storm shots in the 2024 season
}
\keyword{datasets}
