% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Writing - original draft}
\alias{Writing - original draft}
\title{CRediT taxonomy: Writing - original draft}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/writing-original-draft/}
}
\usage{
Writing - original draft
}
\description{
CRediT taxonomy: Writing - original draft
}
\keyword{datasets}
