% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Writing - review & editing}
\alias{Writing - review & editing}
\title{CRediT taxonomy: Writing - review & editing}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/writing-review-editing/}
}
\usage{
Writing - review & editing
}
\description{
CRediT taxonomy: Writing - review & editing
}
\keyword{datasets}
