% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay.R
\name{Assay}
\alias{Assay}
\title{Assay}
\description{
An assay represents a test performed either on material taken from a
subject or on a whole initial subject, producing qualitative or
quantitative measurements.
}
\details{
An assay groups descriptions of provenance of sample processing for
related tests. Each test typically follows the steps of one particular
experimental workflow described by a particular protocol.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measurement_type}}{An \link{OntologyAnnotation} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification)}

\item{\code{technology_type}}{An \link{OntologyAnnotation} to identify the technology  used to perform the measurement}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab}

\item{\code{other_materials}}{Materials associated with the assay}

\item{\code{samples}}{Samples associated with the assay}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object}

\item{\code{characteristic_categories}}{A list of \link{OntologyAnnotation} used in the annotation of material characteristics in the Assay}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level}

\item{\code{comments}}{Comments associated with instances of this class}

\item{\code{data_files}}{\link{DataFile} objects}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object listing all the ontology sources used}

\item{\code{protocols}}{\link{Protocols} objects}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Assay-new}{\code{Assay$new()}}
\item \href{#method-Assay-check_measurement_type}{\code{Assay$check_measurement_type()}}
\item \href{#method-Assay-set_measurement_type}{\code{Assay$set_measurement_type()}}
\item \href{#method-Assay-check_technology_type}{\code{Assay$check_technology_type()}}
\item \href{#method-Assay-set_technology_type}{\code{Assay$set_technology_type()}}
\item \href{#method-Assay-check_technology_platform}{\code{Assay$check_technology_platform()}}
\item \href{#method-Assay-set_technology_platform}{\code{Assay$set_technology_platform()}}
\item \href{#method-Assay-check_filename}{\code{Assay$check_filename()}}
\item \href{#method-Assay-set_filename}{\code{Assay$set_filename()}}
\item \href{#method-Assay-check_comments}{\code{Assay$check_comments()}}
\item \href{#method-Assay-set_comments}{\code{Assay$set_comments()}}
\item \href{#method-Assay-add_comment}{\code{Assay$add_comment()}}
\item \href{#method-Assay-check_samples}{\code{Assay$check_samples()}}
\item \href{#method-Assay-check_sample_ids}{\code{Assay$check_sample_ids()}}
\item \href{#method-Assay-set_samples}{\code{Assay$set_samples()}}
\item \href{#method-Assay-set_samples_by_id}{\code{Assay$set_samples_by_id()}}
\item \href{#method-Assay-to_list}{\code{Assay$to_list()}}
\item \href{#method-Assay-header_table}{\code{Assay$header_table()}}
\item \href{#method-Assay-to_table}{\code{Assay$to_table()}}
\item \href{#method-Assay-from_list}{\code{Assay$from_list()}}
\item \href{#method-Assay-print}{\code{Assay$print()}}
\item \href{#method-Assay-get_process_order}{\code{Assay$get_process_order()}}
\item \href{#method-Assay-get_process_paths}{\code{Assay$get_process_paths()}}
\item \href{#method-Assay-get_process_io_paths}{\code{Assay$get_process_io_paths()}}
\item \href{#method-Assay-clone}{\code{Assay$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-new"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$new(
  measurement_type = NULL,
  technology_type = NULL,
  technology_platform = character(),
  filename = character(),
  other_materials = NULL,
  samples = NULL,
  characteristic_categories = NULL,
  process_sequence = NULL,
  comments = NULL,
  data_files = NULL,
  ontology_source_references = NULL,
  unit_references = NULL,
  protocols = NULL,
  `@id` = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \link{OntologyAnnotation} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}

\item{\code{technology_type}}{An \link{OntologyAnnotation} to identify the technology  used to perform the measurement.}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE.}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab.}

\item{\code{other_materials}}{Materials associated with the assay.}

\item{\code{samples}}{Samples associated with the assay.}

\item{\code{characteristic_categories}}{A list of \link{OntologyAnnotation} used in the annotation of material characteristics in the Assay.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{data_files}}{\link{DataFile} objects.}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object listing all the ontology sources used.}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object.}

\item{\code{protocols}}{\link{Protocols} objects}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new assay
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_measurement_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_measurement_type}{}}}
\subsection{Method \code{check_measurement_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_measurement_type(measurement_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \link{OntologyAnnotation} to qualify the
endpoint, or  what is being measured
(e.g. gene expression profiling or protein  identification).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks the  measurement_type is an instance of \link{OntologyAnnotation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_measurement_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_measurement_type}{}}}
\subsection{Method \code{set_measurement_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_measurement_type(measurement_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \link{OntologyAnnotation} to qualify the
endpoint, or  what is being measured
(e.g. gene expression profiling or protein  identification).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the measurement type
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_technology_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_technology_type}{}}}
\subsection{Method \code{check_technology_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_technology_type(technology_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_type}}{An \link{OntologyAnnotation} to identify the
technology  used to perform the measurement.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks that technology_type an instance of \link{OntologyAnnotation}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_technology_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_technology_type}{}}}
\subsection{Method \code{set_technology_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_technology_type(technology_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_type}}{An \link{OntologyAnnotation} to identify the
technology  used to perform the measurement.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the technology type
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_technology_platform"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_technology_platform}{}}}
\subsection{Method \code{check_technology_platform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_technology_platform(technology_platform)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_platform}}{The technology_platform of the assay}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the technology_platform of the assay is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_technology_platform"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_technology_platform}{}}}
\subsection{Method \code{set_technology_platform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_technology_platform(technology_platform)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_platform}}{The technology_platform of the assay}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the technology_platform of the assay if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_filename"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_filename}{}}}
\subsection{Method \code{check_filename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_filename(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The filename of the assay}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the filename of the assay is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_filename"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_filename}{}}}
\subsection{Method \code{set_filename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_filename(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The filename of the assay}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the filename of the assay if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_samples}{}}}
\subsection{Method \code{check_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_samples(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{a list of \link{Sample} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that all samples are \link{Sample} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_sample_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_sample_ids}{}}}
\subsection{Method \code{check_sample_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_sample_ids(sample_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_ids}}{a vector of sample IDs}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that all sample ids reference one of the supplied sample objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_samples"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_samples}{}}}
\subsection{Method \code{set_samples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_samples(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{a list of \link{Sample} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set samples if samples is a list of \link{Sample} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_samples_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_samples_by_id}{}}}
\subsection{Method \code{set_samples_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_samples_by_id(sample_ids)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-header_table"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-header_table}{}}}
\subsection{Method \code{header_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$header_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-to_table"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-to_table}{}}}
\subsection{Method \code{to_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$to_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}

\item{\code{recursive}}{recursively parse}

\item{\code{json}}{reading from json}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Assay} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-print"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Assay} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-get_process_order"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-get_process_order}{}}}
\subsection{Method \code{get_process_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$get_process_order()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-get_process_paths"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-get_process_paths}{}}}
\subsection{Method \code{get_process_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$get_process_paths()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-get_process_io_paths"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-get_process_io_paths}{}}}
\subsection{Method \code{get_process_io_paths()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$get_process_io_paths()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
