% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assay.R
\name{Assay}
\alias{Assay}
\title{R6 class for an assay}
\description{
An assay represents a test performed either on material taken from a
subject or on a whole initial subject, producing qualitative or
quantitative measurements.
}
\details{
An assay groups descriptions of provenance of sample processing for
related tests. Each test typically follows the steps of one particular
experimental workflow described by a particular protocol.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measurement_type}}{An \code{[OntologyAnnotation]} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}

\item{\code{technology_type}}{An \code{[OntologyAnnotation]} to identify the technology  used to perform the measurement.}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE.}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab.}

\item{\code{materials}}{Materials associated with the assay, lists of '\code{[Sample]}s' and other_material'.}

\item{\code{units}}{A list of units used in the annotation of material units ? \code{[OntologyAnnotation]} !!.}

\item{\code{characteristic_categories}}{A list of \code{[OntologyAnnotation]} used in the annotation of material characteristics in the Assay.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{graph}}{A graph representation of the assay graph.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Assay-new}{\code{Assay$new()}}
\item \href{#method-Assay-check_measurement_type}{\code{Assay$check_measurement_type()}}
\item \href{#method-Assay-check_technology_type}{\code{Assay$check_technology_type()}}
\item \href{#method-Assay-set_measurement_type}{\code{Assay$set_measurement_type()}}
\item \href{#method-Assay-set_technology_type}{\code{Assay$set_technology_type()}}
\item \href{#method-Assay-to_list}{\code{Assay$to_list()}}
\item \href{#method-Assay-check_comments}{\code{Assay$check_comments()}}
\item \href{#method-Assay-set_comments}{\code{Assay$set_comments()}}
\item \href{#method-Assay-from_list}{\code{Assay$from_list()}}
\item \href{#method-Assay-string}{\code{Assay$string()}}
\item \href{#method-Assay-clone}{\code{Assay$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-new"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$new(
  measurement_type = NULL,
  technology_type = NULL,
  technology_platform = NULL,
  filename = NULL,
  materials = NULL,
  units = NULL,
  characteristic_categories = NULL,
  process_sequence = NULL,
  comments = NULL,
  graph = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \code{[OntologyAnnotation]} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}

\item{\code{technology_type}}{An \code{[OntologyAnnotation]} to identify the technology  used to perform the measurement.}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE.}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab.}

\item{\code{materials}}{Materials associated with the assay, lists of '\code{[Sample]}s' and other_material'.}

\item{\code{units}}{A list of units used in the annotation of material units ? \code{[OntologyAnnotation]} !!.}

\item{\code{characteristic_categories}}{A list of \code{[OntologyAnnotation]} used in the annotation of material characteristics in the Assay.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{graph}}{A graph representation of the assay graph.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new assay
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_measurement_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_measurement_type}{}}}
\subsection{Method \code{check_measurement_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_measurement_type(measurement_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \code{[OntologyAnnotation]} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks the  measurement_type is an instance of \code{[OntologyAnnotation]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_technology_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_technology_type}{}}}
\subsection{Method \code{check_technology_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_technology_type(technology_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_type}}{An \code{[OntologyAnnotation]} to identify the technology  used to perform the measurement.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks that technology_type an instance of \code{[OntologyAnnotation]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_measurement_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_measurement_type}{}}}
\subsection{Method \code{set_measurement_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_measurement_type(measurement_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \code{[OntologyAnnotation]} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the measurement type
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_technology_type"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_technology_type}{}}}
\subsection{Method \code{set_technology_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_technology_type(technology_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{technology_type}}{An \code{[OntologyAnnotation]} to identify the technology  used to perform the measurement.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the technology type
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Assay]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-string"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-string}{}}}
\subsection{Method \code{string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$string()}\if{html}{\out{</div>}}
}

\subsection{Details}{
stringify
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Assay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Assay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
