% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{assay}
\alias{assay}
\title{R6 class for an assay}
\description{
An assay represents a test performed either on material taken from a
subject or on a whole initial subject, producing qualitative or
quantitative measurements.
}
\details{
An assay groups descriptions of provenance of sample processing for
related tests. Each test typically follows the steps of one particular
experimental workflow described by a particular protocol.

Create a new assay

set the measurement type

shiny UI element for picking a measurement type

make an R list convertible to json

stringify
}
\section{Public fields}{

}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{measurement_type}}{An \link{ontology_annotation} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}

\item{\code{technology_type}}{An \link{ontology_annotation} to identify the technology  used to perform the measurement.}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE.}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab.}

\item{\code{materials}}{Materials associated with the assay, lists of '\link{sample}s' and \link{other_material}'.}

\item{\code{units}}{A list of units used in the annotation of material units ? \link{ontology_annotation} !!.}

\item{\code{characteristic_categories}}{A list of \link{ontology_annotation} used in the annotation of material characteristics in the Assay.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{graph}}{A graph representation of the assay graph.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-assay-new}{\code{assay$new()}}
\item \href{#method-assay-set_measurement_type}{\code{assay$set_measurement_type()}}
\item \href{#method-assay-get_measurement_type_input}{\code{assay$get_measurement_type_input()}}
\item \href{#method-assay-to_list}{\code{assay$to_list()}}
\item \href{#method-assay-string}{\code{assay$string()}}
\item \href{#method-assay-clone}{\code{assay$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-new"></a>}}
\if{latex}{\out{\hypertarget{method-assay-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$new(
  measurement_type = NA,
  technology_type = NA,
  technology_platform = NA,
  filename = NA,
  materials = NA,
  units = NA,
  characteristic_categories = NA,
  process_sequence = NA,
  comments = NA,
  graph = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \link{ontology_annotation} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}

\item{\code{technology_type}}{An \link{ontology_annotation} to identify the technology  used to perform the measurement.}

\item{\code{technology_platform}}{Manufacturer and platform name, e.g. Bruker AVANCE.}

\item{\code{filename}}{A field to specify the name of the assay file for compatibility with ISA-Tab.}

\item{\code{materials}}{Materials associated with the assay, lists of '\link{sample}s' and \link{other_material}'.}

\item{\code{units}}{A list of units used in the annotation of material units ? \link{ontology_annotation} !!.}

\item{\code{characteristic_categories}}{A list of \link{ontology_annotation} used in the annotation of material characteristics in the Assay.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the Assay level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{graph}}{A graph representation of the assay graph.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-set_measurement_type"></a>}}
\if{latex}{\out{\hypertarget{method-assay-set_measurement_type}{}}}
\subsection{Method \code{set_measurement_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$set_measurement_type(measurement_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{measurement_type}}{An \link{ontology_annotation} to qualify the endpoint, or  what is being measured (e.g. gene expression profiling or protein  identification).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-get_measurement_type_input"></a>}}
\if{latex}{\out{\hypertarget{method-assay-get_measurement_type_input}{}}}
\subsection{Method \code{get_measurement_type_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$get_measurement_type_input()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a shiny UI element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-assay-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-string"></a>}}
\if{latex}{\out{\hypertarget{method-assay-string}{}}}
\subsection{Method \code{string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$string()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assay-clone"></a>}}
\if{latex}{\out{\hypertarget{method-assay-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{assay$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
