% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate-extensions.R
\name{assert_empty}
\alias{assert_empty}
\title{assert_empty}
\usage{
assert_empty(
  x,
  mode = typeof(x),
  null.ok = FALSE,
  zero.len.string.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode}

\item{null.ok}{allow NULL to be valid}

\item{zero.len.string.ok}{accept zero length strings}
}
\value{
x or throws an error
}
\description{
assert_empty
}
\examples{

mode <- "character"
assert_empty(character(), mode)
assert_empty(character(), mode, null.ok = TRUE)
assert_empty(NULL, mode, null.ok = TRUE)
assert_empty("", mode)
assert_empty("", mode, null.ok = TRUE)
}
