% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avs_stats_2021.R
\docType{data}
\name{avs_stats_2021}
\alias{avs_stats_2021}
\title{NHL's Colorado Avalanche Player Statistics}
\format{
A data frame with 196 rows and 50 variables:
\describe{
  \item{person_id}{NHL API Player ID}
  \item{person_full_name}{Player full name}
  \item{season}{Season}
  \item{season_start}{Season start}
  \item{stat_assists}{Assists}
  \item{stat_goals}{Goals}
  \item{stat_pim}{Penalty minutes}
  \item{stat_games}{Games played}
  \item{stat_penalty_minutes}{Penalty minutes}
  \item{stat_plus_minus}{Plus-minus value}
  \item{stat_points}{Points }
  \item{stat_power_play_goals}{Power play goals}
  \item{stat_game_winning_goals}{Game winning goals}
  \item{stat_short_handed_goals}{Short handed goals}
  \item{stat_shots}{Shots}
  \item{stat_shot_pct}{Shot percentage}
  \item{stat_time_on_ice}{Time on ice}
  \item{stat_hits}{Hits}
  \item{stat_power_play_points}{Power play points}
  \item{stat_power_play_time_on_ice}{Power play time on ice}
  \item{stat_even_time_on_ice}{Even play time on ice}
  \item{stat_face_off_pct}{Face off percentage}
  \item{stat_over_time_goals}{Overtime goals}
  \item{stat_short_handed_points}{Short handed points}
  \item{stat_short_handed_time_on_ice}{Short handed time on ice}
  \item{stat_blocked}{Blocks}
  \item{stat_shifts}{Shifts}
  \item{stat_save_percentage}{Save percentage}
  \item{stat_goal_against_average}{Goals against average}
  \item{stat_ot}{OT}
  \item{stat_shutouts}{Shutouts}
  \item{stat_ties}{Ties}
  \item{stat_wins}{Wins}
  \item{stat_losses}{Losses}
  \item{stat_saves}{Saves}
  \item{stat_power_play_saves}{Power play saves}
  \item{stat_short_handed_saves}{Short handed saves}
  \item{stat_even_saves}{Even play saves}
  \item{stat_short_handed_shots}{Short handed shots}
  \item{stat_even_shots}{Even play shots}
  \item{stat_power_play_shots}{Power play shots}
  \item{stat_games_started}{Games started}
  \item{stat_shots_against}{Shots against}
  \item{stat_goals_against}{Goals against}
  \item{stat_power_play_save_percentage}{Power play save percentage}
  \item{stat_short_handed_save_percentage}{Short handed save percentage}
  \item{stat_even_strength_save_percentage}{Even strength save percentage}
  \item{team_name}{Team name}
  \item{team_link}{Team link}
  \item{team_id}{Team ID}
}
}
\source{
\url{https://www.nhl.com/stats/}
}
\usage{
avs_stats_2021
}
\description{
A dataset containing the statistics for all Colorado Avalanche players on the 2021-2 roster
}
\keyword{datasets}
