% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characteristic.R
\name{Characteristic}
\alias{Characteristic}
\title{Characteristic}
\description{
A Characteristic acts as a qualifying property to a material object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{category}}{The classifier of the type of characteristic being described.}

\item{\code{value}}{The value of this instance of a characteristic as relevant to the attached material.}

\item{\code{unit}}{If applicable, a unit qualifier for the value (if the value is numeric).}

\item{\code{category_references}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}

\item{\code{unit_references}}{an \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Characteristic-new}{\code{Characteristic$new()}}
\item \href{#method-Characteristic-check_category}{\code{Characteristic$check_category()}}
\item \href{#method-Characteristic-set_category}{\code{Characteristic$set_category()}}
\item \href{#method-Characteristic-check_value}{\code{Characteristic$check_value()}}
\item \href{#method-Characteristic-set_value}{\code{Characteristic$set_value()}}
\item \href{#method-Characteristic-check_unit}{\code{Characteristic$check_unit()}}
\item \href{#method-Characteristic-set_unit}{\code{Characteristic$set_unit()}}
\item \href{#method-Characteristic-check_comments}{\code{Characteristic$check_comments()}}
\item \href{#method-Characteristic-set_comments}{\code{Characteristic$set_comments()}}
\item \href{#method-Characteristic-add_comment}{\code{Characteristic$add_comment()}}
\item \href{#method-Characteristic-set_valid_category}{\code{Characteristic$set_valid_category()}}
\item \href{#method-Characteristic-set_valid_unit}{\code{Characteristic$set_valid_unit()}}
\item \href{#method-Characteristic-to_table}{\code{Characteristic$to_table()}}
\item \href{#method-Characteristic-to_list}{\code{Characteristic$to_list()}}
\item \href{#method-Characteristic-from_list}{\code{Characteristic$from_list()}}
\item \href{#method-Characteristic-print}{\code{Characteristic$print()}}
\item \href{#method-Characteristic-clone}{\code{Characteristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$new(
  category = NULL,
  value = NULL,
  unit = NULL,
  category_references = NULL,
  comments = NULL,
  `@id` = character(),
  ontology_source_references = NULL,
  unit_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{The classifier of the type of characteristic being described.}

\item{\code{value}}{The value of this instance of a characteristic as relevant to the attached material.}

\item{\code{unit}}{If applicable, a unit qualifier for the value (if the value is numeric).}

\item{\code{category_references}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}

\item{\code{unit_references}}{an \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{Characteristic} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_category}{}}}
\subsection{Method \code{check_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{an \link{CharacteristicCategory} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that category is an \link{CharacteristicCategory} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_category}{}}}
\subsection{Method \code{set_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{an \link{CharacteristicCategory} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set category if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_value}{}}}
\subsection{Method \code{check_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{a value in the specified units}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that value is numeric
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{a value in the specified units}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the value if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_unit}{}}}
\subsection{Method \code{check_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \link{Unit} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if unit is a \link{Unit} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_unit}{}}}
\subsection{Method \code{set_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \link{Unit} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set unit if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_valid_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_valid_category}{}}}
\subsection{Method \code{set_valid_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_valid_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{a category}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set a category and check that it is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_valid_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_valid_unit}{}}}
\subsection{Method \code{set_valid_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_valid_unit(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list to processed into a valid unit object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the unit as a valid ontology term, validity is currently
determined the presence of a unit in a the reference or the ability
to create a new unit object from the list.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-to_table"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-to_table}{}}}
\subsection{Method \code{to_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$to_table()}\if{html}{\out{</div>}}
}

\subsection{Details}{
generate a tabular representation of the characteristic
}

\subsection{Returns}{
a Tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$to_list(ld = FALSE, recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}

\item{\code{recursive}}{call to_list methods of any objects within this
object (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \link{Characteristic} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Characteristic object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)
#' @details
#' set the uuid of this object
#' @param id a uuid
#' @param suffix a human readable suffix
set_id = function(id = uuid::UUIDgenerate(), suffix = character()) {
private$id <- generate_id(id, suffix)
},}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Characteristic} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-print"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty Prints \link{Characteristic} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
