% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Characteristic.R
\name{Characteristic}
\alias{Characteristic}
\title{Characteristic}
\description{
A Characteristic acts as a qualifying property to a material object.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{category}}{The classifier of the type of characteristic being described.}

\item{\code{value}}{The value of this instance of a characteristic as relevant to the attached material.}

\item{\code{unit}}{If applicable, a unit qualifier for the value (if the value is numeric).}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Characteristic-new}{\code{Characteristic$new()}}
\item \href{#method-Characteristic-check_category}{\code{Characteristic$check_category()}}
\item \href{#method-Characteristic-set_category}{\code{Characteristic$set_category()}}
\item \href{#method-Characteristic-check_value}{\code{Characteristic$check_value()}}
\item \href{#method-Characteristic-set_value}{\code{Characteristic$set_value()}}
\item \href{#method-Characteristic-check_unit}{\code{Characteristic$check_unit()}}
\item \href{#method-Characteristic-set_unit}{\code{Characteristic$set_unit()}}
\item \href{#method-Characteristic-check_comments}{\code{Characteristic$check_comments()}}
\item \href{#method-Characteristic-set_comments}{\code{Characteristic$set_comments()}}
\item \href{#method-Characteristic-to_list}{\code{Characteristic$to_list()}}
\item \href{#method-Characteristic-from_list}{\code{Characteristic$from_list()}}
\item \href{#method-Characteristic-get_id}{\code{Characteristic$get_id()}}
\item \href{#method-Characteristic-clone}{\code{Characteristic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$new(category = NULL, value = NULL, unit = NULL, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{The classifier of the type of characteristic being described.}

\item{\code{value}}{The value of this instance of a characteristic as relevant to the attached material.}

\item{\code{unit}}{If applicable, a unit qualifier for the value (if the value is numeric).}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{[Characteristics]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_category}{}}}
\subsection{Method \code{check_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{an \code{[OntologyAnnotation]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that category is an \code{[OntologyAnnotation]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_category"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_category}{}}}
\subsection{Method \code{set_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_category(category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{an \code{[OntologyAnnotation]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set category if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_value}{}}}
\subsection{Method \code{check_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{a value in the specified units}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that value is numeric
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{a value in the specified units}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the value if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_unit}{}}}
\subsection{Method \code{check_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \code{[Unit]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if unit is a \code{[Unit]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_unit"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_unit}{}}}
\subsection{Method \code{set_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \code{[Unit]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set unit if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$to_list(recursive = FALSE, ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recursive}}{call to_list methods of any objects within this object (default FALSE)}

\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \code{[Characteristic]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Characteristic object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Characteristic]} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Characteristic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
