% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatab-parser.R
\name{clear_na_cols_and_transpose}
\alias{clear_na_cols_and_transpose}
\title{clear_na_cols_and_transpose}
\usage{
clear_na_cols_and_transpose(tbl)
}
\arguments{
\item{tbl}{a tibble}
}
\value{
a tibble
}
\description{
convenience wrapper for doing first removal of columns with all NA values
and transposition of a tibble where first row becomes the column names.
This function also adds the step of removing the first (redundant) word from
the sub-section names, this makes it easier to use the same parser for
fields which are in both Study and Investigation such as people and contacts
}
\examples{
clear_na_cols_and_transpose(tibble::tibble(c(NA, NA), 1:2, c(NA, FALSE)))

}
