% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{comment}
\alias{comment}
\title{R6 class for comments}
\description{
A comment allows arbitrary annotation of all commentable ISA classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{a name for the comment string}

\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-comment-new}{\code{comment$new()}}
\item \href{#method-comment-check_name}{\code{comment$check_name()}}
\item \href{#method-comment-check_value}{\code{comment$check_value()}}
\item \href{#method-comment-set_name}{\code{comment$set_name()}}
\item \href{#method-comment-set_value}{\code{comment$set_value()}}
\item \href{#method-comment-to_list}{\code{comment$to_list()}}
\item \href{#method-comment-from_list}{\code{comment$from_list()}}
\item \href{#method-comment-clone}{\code{comment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-new"></a>}}
\if{latex}{\out{\hypertarget{method-comment-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$new(name = "", value = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}

\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make a new comment object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-comment-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that name is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-check_value"></a>}}
\if{latex}{\out{\hypertarget{method-comment-check_value}{}}}
\subsection{Method \code{check_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$check_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check that value is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-comment-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a name for the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set a valid name
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-comment-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$set_value(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the content of the comment string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set a valid value
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-comment-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
convert object to an R list
}

\subsection{Returns}{
a list of length 1 where name is the name and value the value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-comment-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$from_list(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of length 1 where the name is the name and the value the value of the comment}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate a comment object from an R list
}

\subsection{Returns}{
a list of length 1 where name is the name and value the value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-comment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
