% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commentable.R
\name{Commentable}
\alias{Commentable}
\title{R6 superclass for making classes Commentable}
\description{
R6 superclass for making classes Commentable

R6 superclass for making classes Commentable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comments}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Commentable-new}{\code{Commentable$new()}}
\item \href{#method-Commentable-check_comments}{\code{Commentable$check_comments()}}
\item \href{#method-Commentable-set_comments}{\code{Commentable$set_comments()}}
\item \href{#method-Commentable-get_comments}{\code{Commentable$get_comments()}}
\item \href{#method-Commentable-get_comment_names}{\code{Commentable$get_comment_names()}}
\item \href{#method-Commentable-get_comment_values}{\code{Commentable$get_comment_values()}}
\item \href{#method-Commentable-get_comment}{\code{Commentable$get_comment()}}
\item \href{#method-Commentable-add_comment}{\code{Commentable$add_comment()}}
\item \href{#method-Commentable-to_list}{\code{Commentable$to_list()}}
\item \href{#method-Commentable-to_tidy}{\code{Commentable$to_tidy()}}
\item \href{#method-Commentable-from_list}{\code{Commentable$from_list()}}
\item \href{#method-Commentable-clone}{\code{Commentable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-new"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$new(comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[Comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new \code{[Commentable]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[Comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check comments is a list of \code{[Comment]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of \code{[Comment]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set comments if comments is a list of \code{[Comment]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-get_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-get_comments}{}}}
\subsection{Method \code{get_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$get_comments()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get comments as a list of \code{[Comment]} objects
}

\subsection{Returns}{
comments a list of \code{[Comment]} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-get_comment_names"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-get_comment_names}{}}}
\subsection{Method \code{get_comment_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$get_comment_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the names of all comments
}

\subsection{Returns}{
a character vector of comment names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-get_comment_values"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-get_comment_values}{}}}
\subsection{Method \code{get_comment_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$get_comment_values()}\if{html}{\out{</div>}}
}

\subsection{Details}{
get the values of all comments
}

\subsection{Returns}{
a character vector of comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-get_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-get_comment}{}}}
\subsection{Method \code{get_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$get_comment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of the comment to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
get a specific comment by name as a \code{[Comment]} object
}

\subsection{Returns}{
selected comment a \code{[Comment]} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a \code{[Comment]} object to append to the existing \code{[Commentable]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
append a new \code{[Comment]} object to the existing list of comments
}

\subsection{Returns}{
a character vector of comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$to_list()}\if{html}{\out{</div>}}
}

\subsection{Details}{
convert \code{[Commentable]} object into an R list of comments
}

\subsection{Returns}{
a list where the names are comment names and the values comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-to_tidy"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-to_tidy}{}}}
\subsection{Method \code{to_tidy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$to_tidy()}\if{html}{\out{</div>}}
}

\subsection{Details}{
convert \code{[Commentable]} object into an tibble of comments
}

\subsection{Returns}{
a tibble where the first column is comment names and the second is comment values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$from_list(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a R list where the names are comment names and the values are comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
convert an R list into a object \code{[Commentable]}
}

\subsection{Returns}{
commentable a \code{[Commentable]} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Commentable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Commentable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Commentable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
