% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Conceptualization}
\alias{Conceptualization}
\title{CRediT taxonomy: Conceptualization}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/conceptualization/}
}
\usage{
Conceptualization
}
\description{
CRediT taxonomy: Conceptualization
}
\keyword{datasets}
