% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRediT}
\alias{CRediT}
\title{isa_tab_section_headings}
\format{
named character vector

\link{OntologySource} object
}
\source{
https://github.com/ISA-tools/isa-specs/blob/master/source/isatab.rst
CRediT taxonomy

\url{https://credit.niso.org/}
}
\usage{
CRediT
}
\description{
isa_tab_section_headings
}
\keyword{datasets}
