% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFile.R
\name{DataFile}
\alias{DataFile}
\title{Represents a data file in an experimental graph.}
\description{
Represents a data file in an experimental graph.

Represents a data file in an experimental graph.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{A name/reference for the data file.}

\item{\code{label}}{The data file type, as indicated by a label such as 'Array Data File' or 'Raw Data File'}

\item{\code{generated_from}}{Reference to Sample(s) the DataFile is generated from}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataFile-new}{\code{DataFile$new()}}
\item \href{#method-DataFile-check_generated_from}{\code{DataFile$check_generated_from()}}
\item \href{#method-DataFile-set_generated_from}{\code{DataFile$set_generated_from()}}
\item \href{#method-DataFile-check_comments}{\code{DataFile$check_comments()}}
\item \href{#method-DataFile-set_comments}{\code{DataFile$set_comments()}}
\item \href{#method-DataFile-to_list}{\code{DataFile$to_list()}}
\item \href{#method-DataFile-from_list}{\code{DataFile$from_list()}}
\item \href{#method-DataFile-get_id}{\code{DataFile$get_id()}}
\item \href{#method-DataFile-clone}{\code{DataFile$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$new(
  filename = NULL,
  file_path = character(),
  check_file_exists = FALSE,
  compute_hash = FALSE,
  hash_alog = "md5",
  label = NULL,
  generated_from = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{A name/reference for the data file.}

\item{\code{label}}{The data file type, as indicated by a label such as 'Array Data File' or 'Raw Data File'}

\item{\code{generated_from}}{Reference to Sample(s) the DataFile is generated from}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make a new DataFile object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-check_generated_from"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-check_generated_from}{}}}
\subsection{Method \code{check_generated_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$check_generated_from(generated_from)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generated_from}}{an \code{[Sample]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that generated_from is an \code{[Sample]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-set_generated_from"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-set_generated_from}{}}}
\subsection{Method \code{set_generated_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$set_generated_from(generated_from)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{generated_from}}{an \code{[Sample]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set generated_from if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \code{[DataFile]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Characteristic object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[DataFile]} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
