% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{expect_empty}
\alias{expect_empty}
\title{expect_empty}
\usage{
expect_empty(x, mode, null.ok = FALSE, info = NULL, label = vname(x))
}
\arguments{
\item{x}{object to check}

\item{mode}{vector mode}

\item{null.ok}{allow NULL to be valid}
}
\value{
use with testthat
}
\description{
expect_empty
}
\examples{

mode <- "character"
expect_empty(character(), mode)
expect_empty(character(), mode, null.ok = TRUE)
expect_empty(NULL, mode, null.ok = TRUE)
expect_empty("", mode)
expect_empty("", mode, null.ok = TRUE)

}
