% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorValue.R
\name{FactorValue}
\alias{FactorValue}
\title{R6 class for an experimental factor value}
\description{
R6 class for an experimental factor value

R6 class for an experimental factor value
}
\details{
A factor_value represents the value instance of a \link{StudyFactor}.
}
\examples{

## ------------------------------------------------
## Method `FactorValue$new`
## ------------------------------------------------

\dontrun{
FactorValue$new(
    factor_name = sf,
    value = "agrees with",
    unit = NULL
)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{factor}}{an experimental factor}

\item{\code{value}}{value of a quantity}

\item{\code{unit}}{units in which that quantity is measured}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}

\item{\code{study_factor_references}}{a \link{StudyFactorReferences} object}

\item{\code{unit_references}}{an \link{UnitReferences} object}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FactorValue-new}{\code{FactorValue$new()}}
\item \href{#method-FactorValue-check_unit}{\code{FactorValue$check_unit()}}
\item \href{#method-FactorValue-set_unit}{\code{FactorValue$set_unit()}}
\item \href{#method-FactorValue-check_comments}{\code{FactorValue$check_comments()}}
\item \href{#method-FactorValue-set_comments}{\code{FactorValue$set_comments()}}
\item \href{#method-FactorValue-add_comment}{\code{FactorValue$add_comment()}}
\item \href{#method-FactorValue-set_valid_factor_category}{\code{FactorValue$set_valid_factor_category()}}
\item \href{#method-FactorValue-to_list}{\code{FactorValue$to_list()}}
\item \href{#method-FactorValue-from_list}{\code{FactorValue$from_list()}}
\item \href{#method-FactorValue-print}{\code{FactorValue$print()}}
\item \href{#method-FactorValue-clone}{\code{FactorValue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-new"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$new(
  factor = NULL,
  value = NULL,
  unit = NULL,
  ontology_source_references = NULL,
  study_factor_references = NULL,
  unit_references = NULL,
  comments = NULL,
  `@id` = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor}}{an experimental factor}

\item{\code{value}}{the value of a quantity}

\item{\code{unit}}{the unit of measurement}

\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}

\item{\code{study_factor_references}}{a \link{StudyFactorReferences} object}

\item{\code{unit_references}}{an \link{UnitReferences} object}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new factor value
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
FactorValue$new(
    factor_name = sf,
    value = "agrees with",
    unit = NULL
)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-check_unit"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-check_unit}{}}}
\subsection{Method \code{check_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$check_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \code{[Unit]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if unit is a \code{[Unit]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-set_unit"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-set_unit}{}}}
\subsection{Method \code{set_unit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$set_unit(unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a \code{[Unit]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set unit if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-set_valid_factor_category"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-set_valid_factor_category}{}}}
\subsection{Method \code{set_valid_factor_category()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$set_valid_factor_category(factor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor}}{an experimental factor}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set an experimental factor of a valid category
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{OntologyAnnotation} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-print"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{FactorValue} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FactorValue-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FactorValue-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FactorValue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
