% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Funding_acquisition}
\alias{Funding_acquisition}
\title{CRediT taxonomy: Funding acquisition}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/funding-acquisition/}
}
\usage{
Funding_acquisition
}
\description{
CRediT taxonomy: Funding acquisition
}
\keyword{datasets}
