% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySource.R
\name{identical.OntologySource}
\alias{identical.OntologySource}
\title{identical.OntologySource}
\usage{
identical.OntologySource(x, y)
}
\arguments{
\item{x}{a \code{[OntologySource]} object}

\item{y}{a \code{[OntologySource]} object}
}
\description{
Allows checking for the identity of \code{[OntologySource]} objects
}
