% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigation.R
\name{investigation}
\alias{investigation}
\title{R6 class for investigation}
\description{
Represents an investigation from the ISA model
}
\details{
An investigation maintains metadata about the project context and links to one or more studies. There can only be 1 Investigation in an ISA descriptor.
}
\section{Public fields}{

}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{filename}}{...}

\item{\code{identifier}}{A locally unique identifier or an accession number provided by a repository.}

\item{\code{title}}{A concise name given to the investigation.}

\item{\code{description}}{A textual description of the investigation.}

\item{\code{submission_date}}{date on which the investigation was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the investigation should be released publicly. This should be ISO8601 formatted.}

\item{\code{ontology_source_references}}{\code{[ontology_sources]} to be referenced by \code{[ontology_annotations]} used in this ISA descriptor.}

\item{\code{publications}}{A list of Publications associated with an Investigation. (format?)}

\item{\code{contacts}}{A list of People/contacts associated with an Investigation.}

\item{\code{studies}}{\code{[study]} is the central unit, containing information on the subject under study.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-investigation-new}{\code{investigation$new()}}
\item \href{#method-investigation-clone}{\code{investigation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-investigation-new"></a>}}
\if{latex}{\out{\hypertarget{method-investigation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{investigation$new(
  id = "",
  filename = "",
  identifier = "",
  title = "",
  description = "",
  submission_date = "",
  public_release_date = "",
  ontology_source_references = NULL,
  publications = NULL,
  contacts = NULL,
  studies = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{filename}}{...}

\item{\code{identifier}}{A locally unique identifier or an accession number provided by a repository.}

\item{\code{title}}{A concise name given to the investigation.}

\item{\code{description}}{A textual description of the investigation.}

\item{\code{submission_date}}{date on which the investigation was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the investigation should be released publicly. This should be ISO8601 formatted.}

\item{\code{ontology_source_references}}{ontology_source_references \code{[ontology_sources]} to be referenced by \code{[ontology_annotations]} used in this ISA descriptor.}

\item{\code{publications}}{publications A list of Publications associated with an Investigation. (format?)}

\item{\code{contacts}}{contacts A list of People/contacts associated with an Investigation.}

\item{\code{studies}}{studies \code{[study]} is the central unit, containing information on the subject under study.}

\item{\code{comments}}{comments comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new investigation object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-investigation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-investigation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{investigation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
