% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Material.R
\name{Material}
\alias{Material}
\title{R6 object to represent Materials}
\description{
R6 object to represent Materials

R6 object to represent Materials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{name of the material}

\item{\code{type}}{the type of the material}

\item{\code{characteristics}}{the characteristics of the material in the form of a \link{Characteristics} object}

\item{\code{characteristic_categories}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object.}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Material-new}{\code{Material$new()}}
\item \href{#method-Material-check_name}{\code{Material$check_name()}}
\item \href{#method-Material-set_name}{\code{Material$set_name()}}
\item \href{#method-Material-check_type}{\code{Material$check_type()}}
\item \href{#method-Material-set_type}{\code{Material$set_type()}}
\item \href{#method-Material-check_characteristics}{\code{Material$check_characteristics()}}
\item \href{#method-Material-set_characteristics}{\code{Material$set_characteristics()}}
\item \href{#method-Material-check_comments}{\code{Material$check_comments()}}
\item \href{#method-Material-set_comments}{\code{Material$set_comments()}}
\item \href{#method-Material-add_comment}{\code{Material$add_comment()}}
\item \href{#method-Material-to_list}{\code{Material$to_list()}}
\item \href{#method-Material-from_list}{\code{Material$from_list()}}
\item \href{#method-Material-set_id}{\code{Material$set_id()}}
\item \href{#method-Material-print}{\code{Material$print()}}
\item \href{#method-Material-clone}{\code{Material$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-new"></a>}}
\if{latex}{\out{\hypertarget{method-Material-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$new(
  name = character(),
  type = character(),
  characteristics = NULL,
  characteristic_categories = NULL,
  ontology_source_references = NULL,
  unit_references = NULL,
  comments = NULL,
  `@id` = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}

\item{\code{type}}{the type of the material}

\item{\code{characteristics}}{a list of \link{Characteristic} objects}

\item{\code{characteristic_categories}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object.}

\item{\code{comments}}{comments}

\item{\code{@id}}{identifier}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{Material} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the name of the material is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the name of the material if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_type"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_type}{}}}
\subsection{Method \code{check_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check the the type has a non-zero length
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_type"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_type}{}}}
\subsection{Method \code{set_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the type of the \link{Material}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_characteristics}{}}}
\subsection{Method \code{check_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \link{Characteristic} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check characteristics is a list of \link{Characteristic} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_characteristics}{}}}
\subsection{Method \code{set_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \link{Characteristic} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set characteristics if characteristics is a list of \link{Characteristic} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Material-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Material-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Material-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \link{Material} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Material-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Material object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Material} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-Material-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$set_id(id = uuid::UUIDgenerate(), suffix = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{a uuid}

\item{\code{suffix}}{a human readable suffix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the uuid of this object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-print"></a>}}
\if{latex}{\out{\hypertarget{method-Material-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Materials} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Material-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Material-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Material$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
