% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Methodology}
\alias{Methodology}
\title{CRediT taxonomy: Methodology}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/methodology/}
}
\usage{
Methodology
}
\description{
CRediT taxonomy: Methodology
}
\keyword{datasets}
