% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_annotation.R
\name{ontology_annotation}
\alias{ontology_annotation}
\title{R6 class for an experimental factor value}
\description{
R6 class for an experimental factor value

R6 class for an experimental factor value
}
\details{
\link{ontology_source}.
}
\section{Public fields}{

}

\examples{

## ------------------------------------------------
## Method `ontology_annotation$new`
## ------------------------------------------------

ontology_annotation$new(
    term = "agrees with",
    term_source = CiTO,
    term_accession = "agreesWith"
)

## ------------------------------------------------
## Method `ontology_annotation$to_list`
## ------------------------------------------------

ontology_annotation$to_list()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{term}}{the name of ontology term}

\item{\code{term_source}}{the ontology that is the source of the term represented by an \link{ontology_source} object.}

\item{\code{term_accession}}{the unique identifier of the ontology term}

\item{\code{comments}}{comments}

\item{\code{id}}{unique identifier...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ontology_annotation-new}{\code{ontology_annotation$new()}}
\item \href{#method-ontology_annotation-to_list}{\code{ontology_annotation$to_list()}}
\item \href{#method-ontology_annotation-from_list}{\code{ontology_annotation$from_list()}}
\item \href{#method-ontology_annotation-clone}{\code{ontology_annotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology_annotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ontology_annotation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology_annotation$new(
  term = NULL,
  term_source = NULL,
  term_accession = NULL,
  comments = list(),
  id = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{the name of ontology term}

\item{\code{term_source}}{the ontology that is the source of the term represented by an \link{ontology_source} object.}

\item{\code{term_accession}}{the unique identifier of the ontology term}

\item{\code{comments}}{comments}

\item{\code{id}}{a unique identifier ...}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new factor value
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ontology_annotation$new(
    term = "agrees with",
    term_source = CiTO,
    term_accession = "agreesWith"
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology_annotation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-ontology_annotation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology_annotation$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ontology_annotation$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology_annotation-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-ontology_annotation-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology_annotation$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{ontology_annotation} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ontology_annotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ontology_annotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ontology_annotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
