% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologyAnnotation.R
\name{OntologyAnnotation}
\alias{OntologyAnnotation}
\title{R6 class for an experimental factor value}
\description{
R6 class for an experimental factor value

R6 class for an experimental factor value
}
\details{
\code{[OntologySource]}
}
\examples{

## ------------------------------------------------
## Method `OntologyAnnotation$new`
## ------------------------------------------------

OA <- OntologyAnnotation$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{term}}{the name of ontology term}

\item{\code{term_source}}{the ontology that is the source of the term represented by an \code{[OntologySource]} object.}

\item{\code{term_accession}}{the unique identifier of the ontology term}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OntologyAnnotation-new}{\code{OntologyAnnotation$new()}}
\item \href{#method-OntologyAnnotation-check_term_source}{\code{OntologyAnnotation$check_term_source()}}
\item \href{#method-OntologyAnnotation-check_term}{\code{OntologyAnnotation$check_term()}}
\item \href{#method-OntologyAnnotation-check_term_accession}{\code{OntologyAnnotation$check_term_accession()}}
\item \href{#method-OntologyAnnotation-set_term_source}{\code{OntologyAnnotation$set_term_source()}}
\item \href{#method-OntologyAnnotation-set_term}{\code{OntologyAnnotation$set_term()}}
\item \href{#method-OntologyAnnotation-set_term_accession}{\code{OntologyAnnotation$set_term_accession()}}
\item \href{#method-OntologyAnnotation-set_valid_annotation}{\code{OntologyAnnotation$set_valid_annotation()}}
\item \href{#method-OntologyAnnotation-check_comments}{\code{OntologyAnnotation$check_comments()}}
\item \href{#method-OntologyAnnotation-set_comments}{\code{OntologyAnnotation$set_comments()}}
\item \href{#method-OntologyAnnotation-add_comment}{\code{OntologyAnnotation$add_comment()}}
\item \href{#method-OntologyAnnotation-to_list}{\code{OntologyAnnotation$to_list()}}
\item \href{#method-OntologyAnnotation-from_list}{\code{OntologyAnnotation$from_list()}}
\item \href{#method-OntologyAnnotation-get_id}{\code{OntologyAnnotation$get_id()}}
\item \href{#method-OntologyAnnotation-set_id}{\code{OntologyAnnotation$set_id()}}
\item \href{#method-OntologyAnnotation-print}{\code{OntologyAnnotation$print()}}
\item \href{#method-OntologyAnnotation-clone}{\code{OntologyAnnotation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-new"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$new(
  term = character(),
  term_source = NULL,
  term_accession = character(),
  comments = NULL,
  ontology_source_references = NULL,
  `@id` = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{the name of ontology term}

\item{\code{term_source}}{the ontology that is the source of the term represented by an \code{[OntologySource]} object.}

\item{\code{term_accession}}{the unique identifier of the ontology term}

\item{\code{comments}}{comments}

\item{\code{id}}{a unique identifier ...}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new factor value
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{OA <- OntologyAnnotation$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-check_term_source"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-check_term_source}{}}}
\subsection{Method \code{check_term_source()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$check_term_source(term_source)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_source}}{an \code{[OntologySource]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Checks that the source of ontology terms is an \code{[OntologySource]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-check_term"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-check_term}{}}}
\subsection{Method \code{check_term()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$check_term(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{an ontology term}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Checks that the supplied term is in the list of valid terms from the ontology source object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-check_term_accession"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-check_term_accession}{}}}
\subsection{Method \code{check_term_accession()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$check_term_accession(term_accession)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_accession}}{an accession for an ontology term}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Checks that the supplied term accession is in the the list of valid accession terms from the ontology source object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_term_source"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_term_source}{}}}
\subsection{Method \code{set_term_source()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_term_source(term_source)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_source}}{an \code{[OntologySource]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the value of term_source if it passes the checks
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_term"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_term}{}}}
\subsection{Method \code{set_term()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_term(term)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{an ontology term}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the term and the term accession corresponding to that term if the term passes validity checks
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_term_accession"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_term_accession}{}}}
\subsection{Method \code{set_term_accession()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_term_accession(term_accession)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term_accession}}{an accession for an ontology term}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the term accession and the term corresponding to that accession if the accession passes validity checks
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_valid_annotation"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_valid_annotation}{}}}
\subsection{Method \code{set_valid_annotation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_valid_annotation(term, term_accession, term_source_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$to_list(ld = FALSE, recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}

\item{\code{recursive}}{call to_list methods of any objects within this object (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[OntologyAnnotation]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$set_id(id = uuid::UUIDgenerate(), suffix = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{a uuid}

\item{\code{suffix}}{a human readable suffix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the uuid of this object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-print"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologyAnnotation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OntologyAnnotation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologyAnnotation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
