% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySource.R
\name{OntologySource}
\alias{OntologySource}
\title{R6 class for the source of ontology terms}
\description{
An \code{[OntologySource]} describes the resource from which the value of an \code{[OntologyAnnotation]} is derived from.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the source of a term; i.e. the source controlled vocabulary or ontology.}

\item{\code{file}}{A file name or a URI of an official resource.}

\item{\code{file_type}}{the file type, parsable by \code{rdflib::rdf_parse()}}

\item{\code{url}}{...}

\item{\code{version}}{vesion of the ontology}

\item{\code{description}}{free text description of this source of ontology terms}

\item{\code{get_terms_list}}{function to generate valid terms for file}

\item{\code{terms_list}}{list of valid terms}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OntologySource-new}{\code{OntologySource$new()}}
\item \href{#method-OntologySource-to_list}{\code{OntologySource$to_list()}}
\item \href{#method-OntologySource-from_list}{\code{OntologySource$from_list()}}
\item \href{#method-OntologySource-clone}{\code{OntologySource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-new"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$new(
  name = "",
  file = "",
  file_type = "",
  url = "",
  version = "",
  description = "",
  get_terms_list = NULL,
  terms_list = NULL,
  comments = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the source of a term; i.e. the source controlled vocabulary or ontology.}

\item{\code{file}}{A file name or a URI of an official resource.}

\item{\code{file_type}}{the file type, parsable by \code{rdflib::rdf_parse()}}

\item{\code{url}}{...}

\item{\code{version}}{vesion of the ontology}

\item{\code{description}}{free text description of this source of ontology terms}

\item{\code{get_terms_list}}{function to generate valid terms for file}

\item{\code{terms_list}}{list of valid terms}

\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{OntologySource} object to act as source for ontology terms
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{(default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[OntologySource]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
