% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntologySourceReferences.R
\name{OntologySourceReferences}
\alias{OntologySourceReferences}
\title{R6 class for OntologySourceReferences}
\description{
R6 class for OntologySourceReferences

R6 class for OntologySourceReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ontology_source_references}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OntologySourceReferences-new}{\code{OntologySourceReferences$new()}}
\item \href{#method-OntologySourceReferences-from_list}{\code{OntologySourceReferences$from_list()}}
\item \href{#method-OntologySourceReferences-check_ontology_sources}{\code{OntologySourceReferences$check_ontology_sources()}}
\item \href{#method-OntologySourceReferences-add_ontology_sources}{\code{OntologySourceReferences$add_ontology_sources()}}
\item \href{#method-OntologySourceReferences-get_ontology_source_names}{\code{OntologySourceReferences$get_ontology_source_names()}}
\item \href{#method-OntologySourceReferences-get_ontology_source_provision}{\code{OntologySourceReferences$get_ontology_source_provision()}}
\item \href{#method-OntologySourceReferences-clone}{\code{OntologySourceReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$new(ontology_source_references = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$from_list(lst, explicitly_provided = logical())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an ontology source object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{OntologySourceReferences} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-check_ontology_sources"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-check_ontology_sources}{}}}
\subsection{Method \code{check_ontology_sources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$check_ontology_sources(ontology_sources)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology_sources}}{a list of \link{OntologySource} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check ontology_sources is a list of \link{OntologySource} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-add_ontology_sources"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-add_ontology_sources}{}}}
\subsection{Method \code{add_ontology_sources()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$add_ontology_sources(ontology_sources)}\if{html}{\out{</div>}}
}

\subsection{Details}{
Add ontology sources to the
}

\subsection{Returns}{
a vector of ontology source names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-get_ontology_source_names"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-get_ontology_source_names}{}}}
\subsection{Method \code{get_ontology_source_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$get_ontology_source_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the names of the Ontology Sources Used in this Investigation
}

\subsection{Returns}{
a vector of ontology source names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-get_ontology_source_provision"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-get_ontology_source_provision}{}}}
\subsection{Method \code{get_ontology_source_provision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$get_ontology_source_provision()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the provision status of Ontology Sources, where they explicitly
listed in the Investigation or not?
}

\subsection{Returns}{
a vector of ontology source names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OntologySourceReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OntologySourceReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OntologySourceReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
