% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organisation.R
\name{Organisation}
\alias{Organisation}
\title{Organisation}
\description{
Organisation

Organisation
}
\examples{

## ------------------------------------------------
## Method `Organisation$to_list`
## ------------------------------------------------

Organisation$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Organisation-new}{\code{Organisation$new()}}
\item \href{#method-Organisation-check_comments}{\code{Organisation$check_comments()}}
\item \href{#method-Organisation-set_comments}{\code{Organisation$set_comments()}}
\item \href{#method-Organisation-add_comment}{\code{Organisation$add_comment()}}
\item \href{#method-Organisation-to_list}{\code{Organisation$to_list()}}
\item \href{#method-Organisation-from_list}{\code{Organisation$from_list()}}
\item \href{#method-Organisation-get_id}{\code{Organisation$get_id()}}
\item \href{#method-Organisation-set_id}{\code{Organisation$set_id()}}
\item \href{#method-Organisation-clone}{\code{Organisation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$new(comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make a new Organisation object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Organisation$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an \code{[Organisation]} object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Organisation]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$set_id(id = uuid::UUIDgenerate(), suffix = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{a uuid}

\item{\code{suffix}}{a human readable suffix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the uuid of this object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organisation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Organisation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organisation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
