% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\name{Person}
\alias{Person}
\title{R6 class for Person}
\description{
Represents a person
}
\examples{

## ------------------------------------------------
## Method `Person$to_list`
## ------------------------------------------------

Person$new()
}
\section{Super class}{
\code{\link[isar:OntologySourceReferences]{isar::OntologySourceReferences}} -> \code{Person}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Person-new}{\code{Person$new()}}
\item \href{#method-Person-check_email}{\code{Person$check_email()}}
\item \href{#method-Person-set_email}{\code{Person$set_email()}}
\item \href{#method-Person-check_orcid}{\code{Person$check_orcid()}}
\item \href{#method-Person-set_orcid}{\code{Person$set_orcid()}}
\item \href{#method-Person-check_comments}{\code{Person$check_comments()}}
\item \href{#method-Person-set_comments}{\code{Person$set_comments()}}
\item \href{#method-Person-add_comment}{\code{Person$add_comment()}}
\item \href{#method-Person-to_list}{\code{Person$to_list()}}
\item \href{#method-Person-from_list}{\code{Person$from_list()}}
\item \href{#method-Person-get_id}{\code{Person$get_id()}}
\item \href{#method-Person-set_id}{\code{Person$set_id()}}
\item \href{#method-Person-print}{\code{Person$print()}}
\item \href{#method-Person-get_full_name}{\code{Person$get_full_name()}}
\item \href{#method-Person-clone}{\code{Person$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="isar" data-topic="OntologySourceReferences" data-id="add_ontology_sources"><a href='../../isar/html/OntologySourceReferences.html#method-OntologySourceReferences-add_ontology_sources'><code>isar::OntologySourceReferences$add_ontology_sources()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="OntologySourceReferences" data-id="check_ontology_sources"><a href='../../isar/html/OntologySourceReferences.html#method-OntologySourceReferences-check_ontology_sources'><code>isar::OntologySourceReferences$check_ontology_sources()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="OntologySourceReferences" data-id="get_ontology_source_names"><a href='../../isar/html/OntologySourceReferences.html#method-OntologySourceReferences-get_ontology_source_names'><code>isar::OntologySourceReferences$get_ontology_source_names()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-new"></a>}}
\if{latex}{\out{\hypertarget{method-Person-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$new(
  last_name = character(),
  first_name = character(),
  mid_initials = character(),
  email = character(),
  phone = character(),
  fax = character(),
  address = character(),
  affiliation = character(),
  orcid = character(),
  roles = NULL,
  comments = NULL,
  `@id` = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{last_name}}{The last name of a person.}

\item{\code{first_name}}{The first name of a person.}

\item{\code{mid_initials}}{The middle initials of a person.}

\item{\code{email}}{The email address of a person.}

\item{\code{phone}}{The telephone number.}

\item{\code{fax}}{The fax number.}

\item{\code{address}}{The address of a person.}

\item{\code{affiliation}}{The organization affiliation for a person.}

\item{\code{orcid}}{Open Researcher and Contributor ID https://orcid.org/}

\item{\code{roles}}{A list of role(s) performed by this person. Roles reported here need not correspond to roles held withing their affiliated organization.}

\item{\code{comments}}{comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
person
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-check_email"></a>}}
\if{latex}{\out{\hypertarget{method-Person-check_email}{}}}
\subsection{Method \code{check_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$check_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if email address is valid

NB the regex used to validate emails is ... difficult to represent in roxygen syntax
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_email"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_email}{}}}
\subsection{Method \code{set_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{an email address}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the email address field
email addresses ar e first checked for validity by \code{[check_email]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-check_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-Person-check_orcid}{}}}
\subsection{Method \code{check_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$check_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID string}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check if the ORCID is valid i.e. 4 lots of 4 digits separated by hyphens
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_orcid"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_orcid}{}}}
\subsection{Method \code{set_orcid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_orcid(orcid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orcid}}{an ORCID}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
sets the orcid field
ORCID's are first checked for validity by \code{[check_orcid]}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Person-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Person-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Person-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Person$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Person-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$from_list(lst, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an \code{[Person]} object serialized to a list}

\item{\code{json}}{default TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Person]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Person-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-Person-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$set_id(id = uuid::UUIDgenerate(), suffix = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{a uuid}

\item{\code{suffix}}{a human readable suffix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the uuid of this object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-print"></a>}}
\if{latex}{\out{\hypertarget{method-Person-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-get_full_name"></a>}}
\if{latex}{\out{\hypertarget{method-Person-get_full_name}{}}}
\subsection{Method \code{get_full_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$get_full_name()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Combine first name, middle initials, & last name
}

\subsection{Returns}{
string of full name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Person-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Person-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
