% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Project_administration}
\alias{Project_administration}
\title{CRediT taxonomy: Project administration}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/project-administration/}
}
\usage{
Project_administration
}
\description{
CRediT taxonomy: Project administration
}
\keyword{datasets}
