% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protocol.R
\name{Protocol}
\alias{Protocol}
\title{An R6 object a Protocol}
\description{
An R6 object a Protocol

An R6 object a Protocol
}
\details{
An experimental Protocol used in the study.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the protocol used}

\item{\code{protocol_type}}{Term to classify the protocol.}

\item{\code{description}}{A free-text description of the protocol.}

\item{\code{uri}}{Pointer to protocol resources externally that can be accessed by their Uniform Resource Identifier (URI).}

\item{\code{version}}{An identifier for the version to ensure protocol tracking.}

\item{\code{parameters}}{A list of ProtocolParameter describing the list of required to execute the protocol.}

\item{\code{components}}{A list of \link{OntologyAnnotation} describing a protocol's components; e.g. instrument names, software names, and reagents names.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{@id}}{identifier}

\item{\code{origin}}{the id of source assay from which the protocol was generated}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{protocol_references}}{a \link{ProtocolReferences} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Protocol-new}{\code{Protocol$new()}}
\item \href{#method-Protocol-set_ontology_source_references}{\code{Protocol$set_ontology_source_references()}}
\item \href{#method-Protocol-set_protocol_references}{\code{Protocol$set_protocol_references()}}
\item \href{#method-Protocol-check_protocol_references}{\code{Protocol$check_protocol_references()}}
\item \href{#method-Protocol-check_name}{\code{Protocol$check_name()}}
\item \href{#method-Protocol-set_name}{\code{Protocol$set_name()}}
\item \href{#method-Protocol-check_description}{\code{Protocol$check_description()}}
\item \href{#method-Protocol-set_description}{\code{Protocol$set_description()}}
\item \href{#method-Protocol-check_uri}{\code{Protocol$check_uri()}}
\item \href{#method-Protocol-set_uri}{\code{Protocol$set_uri()}}
\item \href{#method-Protocol-check_version}{\code{Protocol$check_version()}}
\item \href{#method-Protocol-set_version}{\code{Protocol$set_version()}}
\item \href{#method-Protocol-check_protocol_type}{\code{Protocol$check_protocol_type()}}
\item \href{#method-Protocol-set_protocol_type}{\code{Protocol$set_protocol_type()}}
\item \href{#method-Protocol-check_parameters}{\code{Protocol$check_parameters()}}
\item \href{#method-Protocol-set_parameters}{\code{Protocol$set_parameters()}}
\item \href{#method-Protocol-add_parameters}{\code{Protocol$add_parameters()}}
\item \href{#method-Protocol-check_components}{\code{Protocol$check_components()}}
\item \href{#method-Protocol-set_components}{\code{Protocol$set_components()}}
\item \href{#method-Protocol-check_comments}{\code{Protocol$check_comments()}}
\item \href{#method-Protocol-set_comments}{\code{Protocol$set_comments()}}
\item \href{#method-Protocol-add_comment}{\code{Protocol$add_comment()}}
\item \href{#method-Protocol-header_table}{\code{Protocol$header_table()}}
\item \href{#method-Protocol-to_list}{\code{Protocol$to_list()}}
\item \href{#method-Protocol-from_list}{\code{Protocol$from_list()}}
\item \href{#method-Protocol-print}{\code{Protocol$print()}}
\item \href{#method-Protocol-clone}{\code{Protocol$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-new"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$new(
  name = character(),
  protocol_type = NULL,
  description = character(),
  uri = character(),
  version = character(),
  parameters = NULL,
  components = NULL,
  comments = NULL,
  `@id` = character(),
  origin = character(),
  ontology_source_references = NULL,
  protocol_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the protocol used}

\item{\code{protocol_type}}{Term to classify the protocol.}

\item{\code{description}}{A free-text description of the protocol.}

\item{\code{uri}}{Pointer to protocol resources externally that can be accessed by their Uniform Resource Identifier (URI).}

\item{\code{version}}{An identifier for the version to ensure protocol tracking.}

\item{\code{parameters}}{A list of \link{ProtocolParameter}s describing the list of to execute the protocol.}

\item{\code{components}}{A list of \link{OntologyAnnotation} describing a protocol's components; e.g. instrument names, software names, and reagents names.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{@id}}{identifier}

\item{\code{origin}}{the id of source assay from which the protocol was generated}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{protocol_references}}{a \link{ProtocolReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new Protocol object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_ontology_source_references"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_ontology_source_references}{}}}
\subsection{Method \code{set_ontology_source_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_ontology_source_references(
  ontology_source_references,
  null.action
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontology_source_references}}{an \link{OntologySourceReferences} object}

\item{\code{null.action}}{how to handle NULLs:
\itemize{
\item "error" throw an error
\item "passthrough" set to NULL
\item "create" set to an empty  \link{OntologySourceReferences} object
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
specify the ontology source references for the \link{Protocol}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_protocol_references"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_protocol_references}{}}}
\subsection{Method \code{set_protocol_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_protocol_references(protocol_references, null.action)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_references}}{an \link{ProtocolReferences} object}

\item{\code{null.action}}{how to handle NULLs:
\itemize{
\item "error" throw an error
\item "passthrough" set to NULL
\item "create" set to an empty  \link{ProtocolReferences} object
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
specify the protocol references for the \link{Protocol}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_protocol_references"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_protocol_references}{}}}
\subsection{Method \code{check_protocol_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_protocol_references(protocol_references)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_references}}{something you want to check is an
\link{ProtocolReferences} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
returns TRUE if protocol_references is a \link{ProtocolReferences}
object and throws an error if it is not
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that name is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set name if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_description"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_description}{}}}
\subsection{Method \code{check_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{the description of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that description is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{the description of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set description if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_uri"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_uri}{}}}
\subsection{Method \code{check_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_uri(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{the uri of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that uri is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_uri"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_uri}{}}}
\subsection{Method \code{set_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_uri(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{the uri of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set uri if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_version"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_version}{}}}
\subsection{Method \code{check_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{the version of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that version is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_version"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_version}{}}}
\subsection{Method \code{set_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{the version of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set version if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_protocol_type"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_protocol_type}{}}}
\subsection{Method \code{check_protocol_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_protocol_type(protocol_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_type}}{an \link{OntologyAnnotation} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that protocol_type is an \link{OntologyAnnotation} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_protocol_type"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_protocol_type}{}}}
\subsection{Method \code{set_protocol_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_protocol_type(protocol_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_type}}{an \link{OntologyAnnotation} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set protocol_type if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_parameters}{}}}
\subsection{Method \code{check_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_parameters(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{a \link{ProtocolParameter} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set parameters if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_parameters}{}}}
\subsection{Method \code{set_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_parameters(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{an \link{ProtocolParameter} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set parameters if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-add_parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-add_parameters}{}}}
\subsection{Method \code{add_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$add_parameters(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{an \link{ProtocolParameter} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add parameters if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_components"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_components}{}}}
\subsection{Method \code{check_components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_components(components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{a list of \link{OntologyAnnotation} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check components is a list of \link{OntologyAnnotation} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_components"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_components}{}}}
\subsection{Method \code{set_components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_components(components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{a list of \link{Study} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set components if components is a list of \link{Study} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-header_table"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-header_table}{}}}
\subsection{Method \code{header_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$header_table()}\if{html}{\out{</div>}}
}

\subsection{Details}{
generate a tabular summary of the protocol for inclusion in
the investigation file
}

\subsection{Returns}{
a Tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$to_list(ld = FALSE, recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}

\item{\code{recursive}}{use the \code{from_list()} method on list items that are also isar objects (default = TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \link{Protocol} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Protocol object serialized to a list}

\item{\code{recursive}}{use the \code{from_list()} method on list items that are also isar objects (default = TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Protocol} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-print"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty Prints \link{Protocol} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Protocol-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Protocol-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
