% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protocol.R
\name{Protocol}
\alias{Protocol}
\title{An R6 object a Protocol}
\description{
An R6 object a Protocol

An R6 object a Protocol
}
\details{
An experimental Protocol used in the study.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the protocol used}

\item{\code{protocol_type}}{Term to classify the protocol.}

\item{\code{description}}{A free-text description of the protocol.}

\item{\code{uri}}{Pointer to protocol resources externally that can be accessed by their Uniform Resource Identifier (URI).}

\item{\code{version}}{An identifier for the version to ensure protocol tracking.}

\item{\code{parameters}}{A list of ProtocolParameter describing the list of required to execute the protocol.}

\item{\code{components}}{A list of \code{[OntologyAnnotation]} describing a protocol's components; e.g. instrument names, software names, and reagents names.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{Protocol$new()}}
\item \href{#method-NA-check_name}{\code{Protocol$check_name()}}
\item \href{#method-NA-set_name}{\code{Protocol$set_name()}}
\item \href{#method-NA-check_description}{\code{Protocol$check_description()}}
\item \href{#method-NA-set_description}{\code{Protocol$set_description()}}
\item \href{#method-NA-check_uri}{\code{Protocol$check_uri()}}
\item \href{#method-NA-set_uri}{\code{Protocol$set_uri()}}
\item \href{#method-NA-check_version}{\code{Protocol$check_version()}}
\item \href{#method-NA-set_version}{\code{Protocol$set_version()}}
\item \href{#method-NA-check_protocol_type}{\code{Protocol$check_protocol_type()}}
\item \href{#method-NA-set_protocol_type}{\code{Protocol$set_protocol_type()}}
\item \href{#method-NA-check_components}{\code{Protocol$check_components()}}
\item \href{#method-NA-set_components}{\code{Protocol$set_components()}}
\item \href{#method-NA-check_comments}{\code{Protocol$check_comments()}}
\item \href{#method-NA-set_comments}{\code{Protocol$set_comments()}}
\item \href{#method-NA-to_list}{\code{Protocol$to_list()}}
\item \href{#method-NA-from_list}{\code{Protocol$from_list()}}
\item \href{#method-NA-get_id}{\code{Protocol$get_id()}}
\item \href{#method-unknown-clone}{\code{Protocol$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$new(
  name = character(),
  protocol_type = NULL,
  description = character(),
  uri = character(),
  version = character(),
  parameters = NULL,
  components = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the protocol used}

\item{\code{protocol_type}}{Term to classify the protocol.}

\item{\code{description}}{A free-text description of the protocol.}

\item{\code{uri}}{Pointer to protocol resources externally that can be accessed by their Uniform Resource Identifier (URI).}

\item{\code{version}}{An identifier for the version to ensure protocol tracking.}

\item{\code{parameters}}{A list of ProtocolParameter describing the list of to execute the protocol.}

\item{\code{components}}{A list of \code{[OntologyAnnotation]} describing a protocol's components; e.g. instrument names, software names, and reagents names.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new Protocol object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that name is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set name if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_description"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_description}{}}}
\subsection{Method \code{check_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{the description of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that description is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{the description of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set description if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_uri"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_uri}{}}}
\subsection{Method \code{check_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_uri(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{the uri of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that uri is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_uri"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_uri}{}}}
\subsection{Method \code{set_uri()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_uri(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{the uri of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set uri if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_version"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_version}{}}}
\subsection{Method \code{check_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{the version of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that version is a single string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_version"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_version}{}}}
\subsection{Method \code{set_version()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_version(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{the version of the protocol}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set version if it passes validation
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_protocol_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_protocol_type}{}}}
\subsection{Method \code{check_protocol_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_protocol_type(protocol_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_type}}{an \code{[OntologyAnnotation]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check that protocol_type is an \code{[OntologyAnnotation]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_protocol_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_protocol_type}{}}}
\subsection{Method \code{set_protocol_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_protocol_type(protocol_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{protocol_type}}{an \code{[OntologyAnnotation]} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set protocol_type if input is valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_components"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_components}{}}}
\subsection{Method \code{check_components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_components(components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{a list of \code{[OntologyAnnotation]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check components is a list of \code{[OntologyAnnotation]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_components"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_components}{}}}
\subsection{Method \code{set_components()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_components(components)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{components}}{a list of \code{[Study]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set components if components is a list of \code{[Study]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \code{[Process]} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-NA-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Protocol object serialized to a list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[Protocol]} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Protocol$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
