% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{R6 class for sample}
\description{
Represents a sample material in an experimental graph.
}
\section{Public fields}{

}

\examples{

## ------------------------------------------------
## Method `sample$to_list`
## ------------------------------------------------

sample$to_list()
}
\section{Super class}{
\code{\link[base:NA]{base::NA}} -> \code{sample}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[factor_values]} used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sample-new}{\code{sample$new()}}
\item \href{#method-sample-check_name}{\code{sample$check_name()}}
\item \href{#method-sample-check_factor_values}{\code{sample$check_factor_values()}}
\item \href{#method-sample-set_name}{\code{sample$set_name()}}
\item \href{#method-sample-set_factor_values}{\code{sample$set_factor_values()}}
\item \href{#method-sample-to_list}{\code{sample$to_list()}}
\item \href{#method-sample-from_list}{\code{sample$from_list()}}
\item \href{#method-sample-clone}{\code{sample$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-new"></a>}}
\if{latex}{\out{\hypertarget{method-sample-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$new(
  id = super$set_id(),
  name = "",
  factor_values = NULL,
  characteristics = NULL,
  derives_from = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[factor_value]} objects used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new instance of sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-sample-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{sample name to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
validates the name field is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-check_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-sample-check_factor_values}{}}}
\subsection{Method \code{check_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$check_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values to be used in a \code{[sample]} object
A list of \code{[factor_value]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
validates the factor_values field is a list of \code{[factor_value]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-sample-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the sample}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the name
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-set_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-sample-set_factor_values}{}}}
\subsection{Method \code{set_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$set_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values used in the sample
A list of \code{[factor_value]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the factor values used in the sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-sample-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{sample$to_list()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-sample-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a \code{[sample]} factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[sample]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
