% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{Sample}
\alias{Sample}
\title{R6 class for sample}
\description{
Represents a sample material in an experimental graph.
}
\examples{

## ------------------------------------------------
## Method `Sample$to_list`
## ------------------------------------------------

Sample$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[FactorValue]}s used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sample-new}{\code{Sample$new()}}
\item \href{#method-Sample-check_factor_values}{\code{Sample$check_factor_values()}}
\item \href{#method-Sample-set_factor_values}{\code{Sample$set_factor_values()}}
\item \href{#method-Sample-to_list}{\code{Sample$to_list()}}
\item \href{#method-Sample-from_list}{\code{Sample$from_list()}}
\item \href{#method-Sample-print}{\code{Sample$print()}}
\item \href{#method-Sample-clone}{\code{Sample$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$new(
  name = character(),
  factor_values = NULL,
  characteristics = NULL,
  derives_from = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[FactorValue]} objects used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new instance of sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-check_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-check_factor_values}{}}}
\subsection{Method \code{check_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$check_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values to be used in a \code{[Sample]} object
A list of \code{[FactorValue]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
validates the factor_values field is a list of \code{[FactorValue]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-set_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-set_factor_values}{}}}
\subsection{Method \code{set_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$set_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values used in the sample
A list of \code{[FactorValue]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the factor values used in the sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Sample$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a \code{[Sample]} factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[sample]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-print"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
