% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample.R
\name{Sample}
\alias{Sample}
\title{R6 class for sample}
\description{
Represents a sample material in an experimental graph.
}
\examples{

## ------------------------------------------------
## Method `Sample$to_list`
## ------------------------------------------------

Sample$new()
}
\section{Super class}{
\code{\link[isar:Commentable]{isar::Commentable}} -> \code{Sample}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[FactorValue]}s used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sample-new}{\code{Sample$new()}}
\item \href{#method-Sample-check_name}{\code{Sample$check_name()}}
\item \href{#method-Sample-check_factor_values}{\code{Sample$check_factor_values()}}
\item \href{#method-Sample-set_name}{\code{Sample$set_name()}}
\item \href{#method-Sample-set_factor_values}{\code{Sample$set_factor_values()}}
\item \href{#method-Sample-to_list}{\code{Sample$to_list()}}
\item \href{#method-Sample-from_list}{\code{Sample$from_list()}}
\item \href{#method-Sample-get_id}{\code{Sample$get_id()}}
\item \href{#method-Sample-clone}{\code{Sample$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="add_comment"><a href='../../isar/html/Commentable.html#method-Commentable-add_comment'><code>isar::Commentable$add_comment()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="check_comments"><a href='../../isar/html/Commentable.html#method-Commentable-check_comments'><code>isar::Commentable$check_comments()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="get_comment"><a href='../../isar/html/Commentable.html#method-Commentable-get_comment'><code>isar::Commentable$get_comment()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="get_comment_names"><a href='../../isar/html/Commentable.html#method-Commentable-get_comment_names'><code>isar::Commentable$get_comment_names()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="get_comment_values"><a href='../../isar/html/Commentable.html#method-Commentable-get_comment_values'><code>isar::Commentable$get_comment_values()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="get_comments"><a href='../../isar/html/Commentable.html#method-Commentable-get_comments'><code>isar::Commentable$get_comments()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="set_comments"><a href='../../isar/html/Commentable.html#method-Commentable-set_comments'><code>isar::Commentable$set_comments()</code></a></li>
<li><span class="pkg-link" data-pkg="isar" data-topic="Commentable" data-id="to_tidy"><a href='../../isar/html/Commentable.html#method-Commentable-to_tidy'><code>isar::Commentable$to_tidy()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$new(
  name = "",
  factor_values = NULL,
  characteristics = NULL,
  derives_from = NULL,
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name/reference for the sample material.}

\item{\code{factor_values}}{A list of \code{[FactorValue]} objects used to qualify the material in terms of study factors/design.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{derives_from}}{A link to the source material that the sample is derived from.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{id}}{...}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new instance of sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$check_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{sample name to validate}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
validates the name field is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-check_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-check_factor_values}{}}}
\subsection{Method \code{check_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$check_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values to be used in a \code{[Sample]} object
A list of \code{[FactorValue]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
validates the factor_values field is a list of \code{[FactorValue]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the sample}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the name
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-set_factor_values"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-set_factor_values}{}}}
\subsection{Method \code{set_factor_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$set_factor_values(factor_values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_values}}{factor values used in the sample
A list of \code{[FactorValue]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets the factor values used in the sample
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Sample$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a \code{[Sample]} factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[sample]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sample$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
