% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Software}
\alias{Software}
\title{CRediT taxonomy: Software}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/software/}
}
\usage{
Software
}
\description{
CRediT taxonomy: Software
}
\keyword{datasets}
