% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{Source}
\alias{Source}
\title{R6 class for a Source}
\description{
R6 class for a Source

R6 class for a Source
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{category_references}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{@id}}{identifier}

\item{\code{unit_references}}{an \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Source-new}{\code{Source$new()}}
\item \href{#method-Source-to_table}{\code{Source$to_table()}}
\item \href{#method-Source-to_list}{\code{Source$to_list()}}
\item \href{#method-Source-from_list}{\code{Source$from_list()}}
\item \href{#method-Source-print}{\code{Source$print()}}
\item \href{#method-Source-clone}{\code{Source$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-new"></a>}}
\if{latex}{\out{\hypertarget{method-Source-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$new(
  name = character(),
  characteristics = NULL,
  comments = NULL,
  category_references = NULL,
  `@id` = character(),
  unit_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{category_references}}{an \link{CharacteristicCategoryReferences} object}

\item{\code{@id}}{identifier}

\item{\code{unit_references}}{an \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-to_table"></a>}}
\if{latex}{\out{\hypertarget{method-Source-to_table}{}}}
\subsection{Method \code{to_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$to_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Source-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)
#' @details

Make \link{Source} from list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Source-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a source object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-print"></a>}}
\if{latex}{\out{\hypertarget{method-Source-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Source} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Source-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
