% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source.R
\name{Source}
\alias{Source}
\title{R6 class for a Source}
\description{
R6 class for a Source

R6 class for a Source
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Source-new}{\code{Source$new()}}
\item \href{#method-Source-check_characteristics}{\code{Source$check_characteristics()}}
\item \href{#method-Source-set_characteristics}{\code{Source$set_characteristics()}}
\item \href{#method-Source-check_comments}{\code{Source$check_comments()}}
\item \href{#method-Source-set_comments}{\code{Source$set_comments()}}
\item \href{#method-Source-to_list}{\code{Source$to_list()}}
\item \href{#method-Source-from_list}{\code{Source$from_list()}}
\item \href{#method-Source-get_id}{\code{Source$get_id()}}
\item \href{#method-Source-clone}{\code{Source$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-new"></a>}}
\if{latex}{\out{\hypertarget{method-Source-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$new(name = character(), characteristics = NULL, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name/reference for the source material.}

\item{\code{characteristics}}{A list of Characteristics used to qualify the material properties.}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-check_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Source-check_characteristics}{}}}
\subsection{Method \code{check_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$check_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \code{[Characteristic]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check characteristics is a list of \code{[Characteristic]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-set_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-Source-set_characteristics}{}}}
\subsection{Method \code{set_characteristics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$set_characteristics(characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{characteristics}}{a list of \code{[Characteristic]} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set characteristics if characteristics is a list of \code{[Characteristic]} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Source-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Source-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Source-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)
#' @details

Make \code{[Source]} from list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
make an R list convertible to json
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Source-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a source object serialized to a list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Source-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Source-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Source-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Source$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
