% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{Study}
\alias{Study}
\title{An R6 Object to represent a study from the ISA model}
\description{
study is the central unit, containing information on the subject under study, its characteristics and any treatments applied.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{A field to specify the name of the Study file corresponding the definition of that Study.}

\item{\code{identifier}}{A unique identifier: either a temporary identifier supplied by users or one generated by a repository or other database.}

\item{\code{@id}}{identifier}

\item{\code{title}}{A concise phrase used to encapsulate the purpose and goal of the study.}

\item{\code{description}}{A textual description of the study, with components such as objective or goals.}

\item{\code{submission_date}}{The date on which the study was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the study should be released publicly. This should be ISO8601 formatted.}

\item{\code{design_descriptors}}{Classifications of the study based on the overall experimental design.}

\item{\code{publications}}{A list of Publications associated with the Study.}

\item{\code{contacts}}{A list of People/contacts associated with the Study.}

\item{\code{factors}}{A factor corresponds to an independent variable manipulated by the experimentalist with the intention to affect biological systems in a way that can be measured by an assay.}

\item{\code{protocols}}{Protocols used within the ISA artifact.}

\item{\code{assays}}{An Assay represents a portion of the experimental design.}

\item{\code{sources}}{Sources associated with the study, is equivalent to materials sources.}

\item{\code{samples}}{samples associated with the study, is equivalent to materials samples.}

\item{\code{other_materials}}{Other Materials associated with the study, is equivalent to materials other_material.}

\item{\code{characteristic_categories}}{Annotations of material characteristics used in the study.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the study level.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object.

#field materials Materials associated with the study, contains lists of 'sources', 'samples' and 'other_material'. DEPRECATED.
#field graph Graph representation of the study graph.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Study-new}{\code{Study$new()}}
\item \href{#method-Study-check_title}{\code{Study$check_title()}}
\item \href{#method-Study-set_title}{\code{Study$set_title()}}
\item \href{#method-Study-check_description}{\code{Study$check_description()}}
\item \href{#method-Study-set_description}{\code{Study$set_description()}}
\item \href{#method-Study-check_public_release_date}{\code{Study$check_public_release_date()}}
\item \href{#method-Study-set_public_release_date}{\code{Study$set_public_release_date()}}
\item \href{#method-Study-check_submission_date}{\code{Study$check_submission_date()}}
\item \href{#method-Study-set_submission_date}{\code{Study$set_submission_date()}}
\item \href{#method-Study-check_contacts}{\code{Study$check_contacts()}}
\item \href{#method-Study-check_publications}{\code{Study$check_publications()}}
\item \href{#method-Study-set_publications}{\code{Study$set_publications()}}
\item \href{#method-Study-set_contacts}{\code{Study$set_contacts()}}
\item \href{#method-Study-check_comments}{\code{Study$check_comments()}}
\item \href{#method-Study-set_comments}{\code{Study$set_comments()}}
\item \href{#method-Study-add_comment}{\code{Study$add_comment()}}
\item \href{#method-Study-to_list}{\code{Study$to_list()}}
\item \href{#method-Study-from_list}{\code{Study$from_list()}}
\item \href{#method-Study-print}{\code{Study$print()}}
\item \href{#method-Study-clone}{\code{Study$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-new"></a>}}
\if{latex}{\out{\hypertarget{method-Study-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(
  filename = "",
  identifier = character(),
  `@id` = character(),
  title = character(),
  description = character(),
  submission_date = NULL,
  public_release_date = NULL,
  contacts = NULL,
  design_descriptors = NULL,
  publications = NULL,
  factors = NULL,
  protocols = NULL,
  assays = NULL,
  sources = NULL,
  samples = NULL,
  process_sequence = NULL,
  other_materials = NULL,
  characteristic_categories = NULL,
  comments = NULL,
  ontology_source_references = NULL,
  unit_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{A field to specify the name of the Study file corresponding the definition of that Study.}

\item{\code{identifier}}{A unique identifier: either a temporary identifier supplied by users or one generated by a repository or other database.}

\item{\code{@id}}{A unique identifier: either a temporary identifier supplied by users or one generated by a repository or other database.}

\item{\code{title}}{A concise phrase used to encapsulate the purpose and goal of the study.}

\item{\code{description}}{A textual description of the study, with components such as objective or goals.}

\item{\code{submission_date}}{The date on which the study was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the study should be released publicly. This should be ISO8601 formatted.}

\item{\code{contacts}}{A list of People/contacts associated with the Study.}

\item{\code{design_descriptors}}{Classifications of the study based on the overall experimental design.}

\item{\code{publications}}{A list of Publications associated with the Study.}

\item{\code{factors}}{A factor corresponds to an independent variable manipulated by the experimentalist with the intention to affect biological systems in a way that can be measured by an assay.}

\item{\code{protocols}}{Protocols used within the ISA artifact.}

\item{\code{assays}}{An Assay represents a portion of the experimental design.}

\item{\code{sources}}{Sources associated with the study, is equivalent to materials sources.}

\item{\code{samples}}{samples associated with the study, is equivalent to materials \link{Sample}.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the study level.}

\item{\code{other_materials}}{Other Materials associated with the study, is equivalent to materials other_material.}

\item{\code{characteristic_categories}}{Annotations of material characteristics used in the study.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor.}

\item{\code{unit_references}}{A list of Units used in the annotation of material units in the study.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new study object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_title"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_title}{}}}
\subsection{Method \code{check_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{The title of the study}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the title of the study is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_title}{}}}
\subsection{Method \code{set_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{The title of the study}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the title of the study if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_description"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_description}{}}}
\subsection{Method \code{check_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the study}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the description of the study is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_description"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_description}{}}}
\subsection{Method \code{set_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_description(description)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{The description of the study}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the description of the study if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_public_release_date"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_public_release_date}{}}}
\subsection{Method \code{check_public_release_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_public_release_date(public_release_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{public_release_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check public_release_date is a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_public_release_date"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_public_release_date}{}}}
\subsection{Method \code{set_public_release_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_public_release_date(public_release_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{public_release_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set public_release_date to a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_submission_date"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_submission_date}{}}}
\subsection{Method \code{check_submission_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_submission_date(submission_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check submission_date is a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_submission_date"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_submission_date}{}}}
\subsection{Method \code{set_submission_date()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_submission_date(submission_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{submission_date}}{a Date Object or ISO8601 formatted data string i.e. YYYY-mm-dd}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set submission_date to a Date object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_contacts}{}}}
\subsection{Method \code{check_contacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_contacts(contacts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contacts}}{a list of \link{Person} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check contacts is a list of \link{Person} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_publications"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_publications}{}}}
\subsection{Method \code{check_publications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_publications(publications)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{publications}}{a list of \link{Publication} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
check publications is a list of \link{Publication} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_publications"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_publications}{}}}
\subsection{Method \code{set_publications()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_publications(publications)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{publications}}{a list of \link{Publication} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check publications is a list of \link{Publication} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_contacts}{}}}
\subsection{Method \code{set_contacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_contacts(contacts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contacts}}{a list of \link{Person} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set contacts if contacts is a list of \link{Person} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Study-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-Study-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-Study-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Study-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \link{Study} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Study-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Material object serialized to a list}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}

\item{\code{json}}{json  (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Material} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-print"></a>}}
\if{latex}{\out{\hypertarget{method-Study-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Study} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Study-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Study-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
