% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{An R6 Object to represent a study from the ISA model}
\description{
study is the central unit, containing information on the subject under study, its characteristics and any treatments applied.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{...}

\item{\code{filename}}{A field to specify the name of the Study file corresponding the definition of that Study.}

\item{\code{identifier}}{A unique identifier: either a temporary identifier supplied by users or one generated by a repository or other database.}

\item{\code{title}}{A concise phrase used to encapsulate the purpose and goal of the study.}

\item{\code{description}}{A textual description of the study, with components such as objective or goals.}

\item{\code{submission_date}}{The date on which the study was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the study should be released publicly. This should be ISO8601 formatted.}

\item{\code{design_descriptors}}{Classifications of the study based on the overall experimental design.}

\item{\code{publications}}{A list of Publications associated with the Study.}

\item{\code{contacts}}{A list of People/contacts associated with the Study.}

\item{\code{factors}}{A factor corresponds to an independent variable manipulated by the experimentalist with the intention to affect biological systems in a way that can be measured by an assay.}

\item{\code{protocols}}{Protocols used within the ISA artifact.}

\item{\code{assays}}{An Assay represents a portion of the experimental design.}

\item{\code{sources}}{Sources associated with the study, is equivalent to materials\link{'sources'}.}

\item{\code{samples}}{samples associated with the study, is equivalent to materials\link{'samples'}.}

\item{\code{other_material}}{Other Materials associated with the study, is equivalent to materials\link{'other_material'}.}

\item{\code{units}}{A list of Units used in the annotation of material units in the study.}

\item{\code{characteristic_categories}}{Annotations of material characteristics used in the study.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the study level.}

\item{\code{comments}}{Comments associated with instances of this class.

#field materials Materials associated with the study, contains lists of 'sources', 'samples' and 'other_material'. DEPRECATED.
#field graph Graph representation of the study graph.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-study-new}{\code{study$new()}}
\item \href{#method-study-clone}{\code{study$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study-new"></a>}}
\if{latex}{\out{\hypertarget{method-study-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study$new(
  id = "",
  filename = "",
  identifier = "",
  title = "",
  description = "",
  submission_date = "",
  public_release_date = "",
  contacts = NULL,
  design_descriptors = NULL,
  publications = NULL,
  factors = NULL,
  protocols = NULL,
  assays = NULL,
  sources = NULL,
  samples = NULL,
  process_sequence = NULL,
  other_material = NULL,
  characteristic_categories = NULL,
  comments = NULL,
  units = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{...}

\item{\code{filename}}{A field to specify the name of the Study file corresponding the definition of that Study.}

\item{\code{identifier}}{A unique identifier: either a temporary identifier supplied by users or one generated by a repository or other database.}

\item{\code{title}}{A concise phrase used to encapsulate the purpose and goal of the study.}

\item{\code{description}}{A textual description of the study, with components such as objective or goals.}

\item{\code{submission_date}}{The date on which the study was reported to the repository. This should be ISO8601 formatted.}

\item{\code{public_release_date}}{The date on which the study should be released publicly. This should be ISO8601 formatted.}

\item{\code{contacts}}{A list of People/contacts associated with the Study.}

\item{\code{design_descriptors}}{Classifications of the study based on the overall experimental design.}

\item{\code{publications}}{A list of Publications associated with the Study.}

\item{\code{factors}}{A factor corresponds to an independent variable manipulated by the experimentalist with the intention to affect biological systems in a way that can be measured by an assay.}

\item{\code{protocols}}{Protocols used within the ISA artifact.}

\item{\code{assays}}{An Assay represents a portion of the experimental design.}

\item{\code{sources}}{Sources associated with the study, is equivalent to materials\link{'sources'}.}

\item{\code{samples}}{samples associated with the study, is equivalent to materials\link{'samples'}.}

\item{\code{process_sequence}}{A list of Process objects representing the experimental graphs at the study level.}

\item{\code{other_material}}{Other Materials associated with the study, is equivalent to materials\link{'other_material'}.}

\item{\code{characteristic_categories}}{Annotations of material characteristics used in the study.}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{units}}{A list of Units used in the annotation of material units in the study.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new study object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study-clone"></a>}}
\if{latex}{\out{\hypertarget{method-study-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
