% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_factor.R
\name{study_factor}
\alias{study_factor}
\title{R6 class for an experimental factors factors}
\description{
R6 class for an experimental factors factors

R6 class for an experimental factors factors
}
\details{
A study_factor corresponds to an independent variable manipulated by the
experimentalist with the intention to affect biological systems in a way
that can be measured by an assay.
}
\section{Public fields}{

\itemize{
\item \code{id}: name of a quantity
\item \code{name}: value of a quantity
\item \code{factor_type}: An ontology source reference of the study factor type
\item \code{comments}: Comments associated with instances of this class.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-study_factor-new}{\code{study_factor$new()}}
\item \href{#method-study_factor-clone}{\code{study_factor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-new"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$new(
  id = "",
  name = "",
  factor_type = ontology_annotation$new(),
  comments = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id:}}{unique identifier...}

\item{\code{name:}}{name of factor}

\item{\code{factor_type:}}{An ontology source reference of the study factor type}

\item{\code{comments:}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Public fields:
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-study_factor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-study_factor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{study_factor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
