% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyFactor.R
\name{StudyFactor}
\alias{StudyFactor}
\title{R6 class for an experimental factor}
\description{
A \code{[StudyFactor]} corresponds to an independent variable manipulated by the
experimentalist with the intention to affect biological systems in a way
that can be measured by an assay.
}
\examples{

## ------------------------------------------------
## Method `StudyFactor$to_list`
## ------------------------------------------------

StudyFactor$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[OntologySource]} reference of the study factor type}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyFactor-new}{\code{StudyFactor$new()}}
\item \href{#method-StudyFactor-to_list}{\code{StudyFactor$to_list()}}
\item \href{#method-StudyFactor-from_list}{\code{StudyFactor$from_list()}}
\item \href{#method-StudyFactor-get_id}{\code{StudyFactor$get_id()}}
\item \href{#method-StudyFactor-clone}{\code{StudyFactor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$new(name = "", factor_type = NULL, comments = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[OntologySource]} reference of the study factor_type}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new study factor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{StudyFactor$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$from_list(lst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a study factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[OntologyAnnotation]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
