% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyFactor.R
\name{StudyFactor}
\alias{StudyFactor}
\title{R6 class for an experimental factor}
\description{
A \code{[StudyFactor]} corresponds to an independent variable manipulated by the
experimentalist with the intention to affect biological systems in a way
that can be measured by an assay.
}
\examples{

## ------------------------------------------------
## Method `StudyFactor$to_list`
## ------------------------------------------------

StudyFactor$new()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[OntologyAnnotation]} reference of the study factor type}

\item{\code{comments}}{Comments associated with instances of this class.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{The name of the factor}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyFactor-new}{\code{StudyFactor$new()}}
\item \href{#method-StudyFactor-check_name}{\code{StudyFactor$check_name()}}
\item \href{#method-StudyFactor-set_name}{\code{StudyFactor$set_name()}}
\item \href{#method-StudyFactor-check_comments}{\code{StudyFactor$check_comments()}}
\item \href{#method-StudyFactor-set_comments}{\code{StudyFactor$set_comments()}}
\item \href{#method-StudyFactor-add_comment}{\code{StudyFactor$add_comment()}}
\item \href{#method-StudyFactor-to_list}{\code{StudyFactor$to_list()}}
\item \href{#method-StudyFactor-from_list}{\code{StudyFactor$from_list()}}
\item \href{#method-StudyFactor-get_id}{\code{StudyFactor$get_id()}}
\item \href{#method-StudyFactor-set_id}{\code{StudyFactor$set_id()}}
\item \href{#method-StudyFactor-print}{\code{StudyFactor$print()}}
\item \href{#method-StudyFactor-clone}{\code{StudyFactor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$new(
  factor_name = character(),
  factor_type = NULL,
  comments = NULL,
  `@id` = character(),
  explicitly_provided = logical(),
  ontology_source_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_name}}{The name of the factor}

\item{\code{factor_type}}{An \code{[OntologyAnnotation]} reference of the study factor_type}

\item{\code{comments}}{Comments associated with instances of this class.}

\item{\code{explicitly_provided}}{Explicitly listed in the study as a factor (logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
create a new study factor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-check_name"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-check_name}{}}}
\subsection{Method \code{check_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$check_name(factor_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if the name of the material is a string
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-set_name}{}}}
\subsection{Method \code{set_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$set_name(factor_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{factor_name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the name of the material if valid
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-check_comments"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-check_comments}{}}}
\subsection{Method \code{check_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$check_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
checks if comments are a named list of character vectors
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-set_comments"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-set_comments}{}}}
\subsection{Method \code{set_comments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$set_comments(comments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comments}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Sets comments if they are in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-add_comment"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-add_comment}{}}}
\subsection{Method \code{add_comment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$add_comment(comment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{a list of comments}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add comment if it is in a valid format
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$to_list(ld = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ld}}{logical json-ld}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
generate an R list representation translatable to JSON
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{StudyFactor$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$from_list(lst, recursive = TRUE, json = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list serialization of a study factor object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \code{[StudyFactor]} from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$get_id()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the uuid of this object
}

\subsection{Returns}{
a uuid
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-set_id}{}}}
\subsection{Method \code{set_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$set_id(id = uuid::UUIDgenerate(), suffix = character())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{a uuid}

\item{\code{suffix}}{a human readable suffix}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
set the uuid of this object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-print"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
