% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyFactorReferences.R
\name{StudyFactorReferences}
\alias{StudyFactorReferences}
\title{R6 Class for StudyFactorReferences}
\description{
R6 Class for StudyFactorReferences

R6 Class for StudyFactorReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{study_factor_references}}{a list \link{StudyFactor} objects}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StudyFactorReferences-new}{\code{StudyFactorReferences$new()}}
\item \href{#method-StudyFactorReferences-check_study_factors}{\code{StudyFactorReferences$check_study_factors()}}
\item \href{#method-StudyFactorReferences-add_study_factors}{\code{StudyFactorReferences$add_study_factors()}}
\item \href{#method-StudyFactorReferences-header_table}{\code{StudyFactorReferences$header_table()}}
\item \href{#method-StudyFactorReferences-to_list}{\code{StudyFactorReferences$to_list()}}
\item \href{#method-StudyFactorReferences-from_list}{\code{StudyFactorReferences$from_list()}}
\item \href{#method-StudyFactorReferences-get_study_factor_names}{\code{StudyFactorReferences$get_study_factor_names()}}
\item \href{#method-StudyFactorReferences-get_study_factor_ids}{\code{StudyFactorReferences$get_study_factor_ids()}}
\item \href{#method-StudyFactorReferences-print}{\code{StudyFactorReferences$print()}}
\item \href{#method-StudyFactorReferences-clone}{\code{StudyFactorReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$new(
  study_factor_references = NULL,
  ontology_source_references = NULL,
  unit_references = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_factor_references}}{a list \link{StudyFactor} objects}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-check_study_factors"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-check_study_factors}{}}}
\subsection{Method \code{check_study_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$check_study_factors(study_factors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_factors}}{a list of \link{StudyFactor} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check if this input is a list of \link{StudyFactor} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-add_study_factors"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-add_study_factors}{}}}
\subsection{Method \code{add_study_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$add_study_factors(study_factors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study_factors}}{a list of \link{StudyFactor} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add new study factors to the study factor reference
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-header_table"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-header_table}{}}}
\subsection{Method \code{header_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$header_table()}\if{html}{\out{</div>}}
}

\subsection{Details}{
generate a tabular summary of the study factors for use in the
isa-tab investigation file.
}

\subsection{Returns}{
a Tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$to_list()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the material}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Generate an R list representation of a \link{StudyFactorReferences} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$from_list(lst, origin = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{a list}

\item{\code{origin}}{id of the source which supplied the study factors}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Populate a \link{StudyFactorReferences} object from a list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-get_study_factor_names"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-get_study_factor_names}{}}}
\subsection{Method \code{get_study_factor_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$get_study_factor_names()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the names of the study factor references
}

\subsection{Returns}{
character vector of study factor reference names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-get_study_factor_ids"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-get_study_factor_ids}{}}}
\subsection{Method \code{get_study_factor_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$get_study_factor_ids()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the @ids of the study factor references
}

\subsection{Returns}{
character vector of study factor reference @ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-print"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{StudyFactorReferences} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StudyFactorReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StudyFactorReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StudyFactorReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
