% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit.R
\name{Unit}
\alias{Unit}
\title{R6 object for specifying a unit of measurement}
\description{
R6 object for specifying a unit of measurement

R6 object for specifying a unit of measurement
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Unit-new}{\code{Unit$new()}}
\item \href{#method-Unit-clone}{\code{Unit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-new"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$new(unit = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
