% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unit.R
\name{Unit}
\alias{Unit}
\title{R6 object for specifying a unit of measurement}
\description{
R6 object for specifying a unit of measurement

R6 object for specifying a unit of measurement
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{unit}}{a unit of measurement}

\item{\code{@id}}{identifier}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}

\item{\code{source}}{the source of the \link{OntologySource} object, was it listed in}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Unit-new}{\code{Unit$new()}}
\item \href{#method-Unit-get_unit_string}{\code{Unit$get_unit_string()}}
\item \href{#method-Unit-to_list}{\code{Unit$to_list()}}
\item \href{#method-Unit-from_list}{\code{Unit$from_list()}}
\item \href{#method-Unit-print}{\code{Unit$print()}}
\item \href{#method-Unit-clone}{\code{Unit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-new"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$new(
  unit = NULL,
  `@id` = character(),
  ontology_source_references = NULL,
  source = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{unit}}{a unit of measurement}

\item{\code{@id}}{identifier}

\item{\code{ontology_source_references}}{\link{OntologySource}s to be referenced by \link{OntologyAnnotation}s used in this ISA descriptor}

\item{\code{source}}{the source of the \link{OntologySource} object, was it listed in}

\item{\code{unit_references}}{A list of units used as a \link{UnitReferences} object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \link{Unit} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-get_unit_string"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-get_unit_string}{}}}
\subsection{Method \code{get_unit_string()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$get_unit_string()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the term associated with the unit as a string
}

\subsection{Returns}{
a character vector of length one with the term of the unit
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$to_list(recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recursive}}{linked data (default FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
An R list representation of a \link{Unit} object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$from_list(lst, source = NA, add = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lst}}{an Unit object serialized to a list}

\item{\code{json}}{json  (default TRUE)}

\item{\code{recursive}}{call to_list methods of any objects within this object (default TRUE)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Make \link{Unit} object from list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-print"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Pretty prints \link{Unit} objects
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Unit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Unit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Unit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
