% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitReferences.R
\name{UnitReferences}
\alias{UnitReferences}
\title{R6 class for UnitReferences}
\description{
R6 class for UnitReferences

R6 class for UnitReferences
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{unit_references}}{...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UnitReferences-new}{\code{UnitReferences$new()}}
\item \href{#method-UnitReferences-check_unit_references}{\code{UnitReferences$check_unit_references()}}
\item \href{#method-UnitReferences-set_unit_references}{\code{UnitReferences$set_unit_references()}}
\item \href{#method-UnitReferences-add_unit_references}{\code{UnitReferences$add_unit_references()}}
\item \href{#method-UnitReferences-get_unit_ids}{\code{UnitReferences$get_unit_ids()}}
\item \href{#method-UnitReferences-from_list}{\code{UnitReferences$from_list()}}
\item \href{#method-UnitReferences-clone}{\code{UnitReferences$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-new"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$new(unit_references = NULL, ontology_source_references = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-check_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-check_unit_references}{}}}
\subsection{Method \code{check_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$check_unit_references(units)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-set_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-set_unit_references}{}}}
\subsection{Method \code{set_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$set_unit_references(units)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-add_unit_references"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-add_unit_references}{}}}
\subsection{Method \code{add_unit_references()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$add_unit_references(units)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-get_unit_ids"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-get_unit_ids}{}}}
\subsection{Method \code{get_unit_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$get_unit_ids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$from_list(lst, source = NA, add = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UnitReferences-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UnitReferences-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UnitReferences$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
