% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Validation}
\alias{Validation}
\title{CRediT taxonomy: Validation}
\format{
\link{OntologyAnnotation} object
}
\source{
\url{https://credit.niso.org/contributor-roles/validation/}
}
\usage{
Validation
}
\description{
CRediT taxonomy: Validation
}
\keyword{datasets}
